% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleDataDescription.R
\docType{data}
\name{centrality_data_harmony}
\alias{centrality_data_harmony}
\title{Example data for plotting a Semantic Centrality Plot.}
\format{
A data frame with 2,146 and 4 variables:
\describe{
  \item{words}{unique words}
  \item{n}{overall word frequency}
  \item{central_semantic_similarity}{cosine semantic similarity to the aggregated word embedding}
  \item{n_percent}{frequency in percent}
}
}
\source{
\url{https://link.springer.com/article/10.1007/s11205-015-0903-z}
}
\usage{
centrality_data_harmony
}
\description{
The dataset is a shortened version of the data sets of Study 1
from Kjell, et al., 2016.
}
\keyword{datasets}
