% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleDataDescription.R
\docType{data}
\name{DP_projections_HILS_SWLS_100}
\alias{DP_projections_HILS_SWLS_100}
\title{Data for plotting a Dot Product Projection Plot.}
\format{
A data frame with 583 rows and 12 variables:
\describe{
  \item{words}{unique words}
  \item{dot.x}{dot product projection on the x-axes}
  \item{p_values_dot.x}{p-value for the word in relation to the x-axes}
  \item{n_g1.x}{frequency of the word in group 1 on the x-axes variable}
  \item{n_g2.x}{frequency of the word in group 2 on the x-axes variable}
  \item{dot.y}{dot product projection on the y-axes}
  \item{p_values_dot.y}{p-value for the word in relation to the y-axes}
  \item{n_g1.y}{frequency of the word in group 1 on the y-axes variable}
  \item{n_g2.y}{frequency of the word in group 2 on the x-axes variable}
  \item{n}{overall word frequency}
  \item{n.percent}{frequency in percent}
  \item{N_participant_responses}{number of participants (as this is needed
   in the analyses)}
}
}
\usage{
DP_projections_HILS_SWLS_100
}
\description{
Tibble is the output from textProjection.
The dataset is a shortened version of the data sets of Study 3-5
from Kjell, Kjell, Garcia and Sikström 2018.
}
\keyword{datasets}
