% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_5_textTrainPredictExamples.R
\name{dynamic_jitter_data}
\alias{dynamic_jitter_data}
\title{Add random noise to variables}
\usage{
dynamic_jitter_data(
  data,
  x_col,
  y_col = NULL,
  jitter_factor = 0.05,
  seed = 2025
)
}
\arguments{
\item{data}{(dataframe) Dataframe including the varaibles that should get jitter.}

\item{x_col}{(string) name of the column that should get jitter.}

\item{y_col}{(string) name of the column that should get jitter.}

\item{jitter_factor}{(numeric) Percentage of jitter to be added.}

\item{seed}{(numeric)}
}
\description{
This function is needed because adding jitter des not work to the ggplot objective
from the topicsScatterLegend() function.
}
