% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleDataDescription.R
\docType{data}
\name{embeddings_from_huggingface2}
\alias{embeddings_from_huggingface2}
\title{Word embeddings from textHuggingFace function}
\format{
A list with word embeddings for harmony words for only contexts. BERT-base embeddings based on
mean aggregation of layer 1 and 2.
\describe{
  \item{tokens}{words}
  \item{layer_number }{layer of the transformer model}
  \item{Dim1:Dim768}{Word embeddings dimensions}
}
}
\source{
\url{https://psyarxiv.com/er6t7/}
}
\usage{
embeddings_from_huggingface2
}
\description{
The dataset is a shortened version of the data sets of Study 3-5
from Kjell, Kjell, Garcia and Sikström 2018.
}
\keyword{datasets}
