% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_0_1_text_initialize.R
\name{find_textrpp}
\alias{find_textrpp}
\title{Find text required python packages}
\usage{
find_textrpp(ask)
}
\arguments{
\item{ask}{logical; if \code{FALSE}, use the first text required python packages installation found;
if \code{TRUE}, list available text required python packages installations and prompt the user
for which to use. If another (e.g. \code{python_executable}) is set, then
this value will always be treated as \code{FALSE}.}
}
\value{
textrpp_python
}
\description{
Locate the user's version of Python for which text required python packages are installed.
}
\keyword{internal}
