% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_0_1_text_initialize.R
\name{find_textrpp_env}
\alias{find_textrpp_env}
\title{Find text required python pacakges env}
\usage{
find_textrpp_env()
}
\description{
check whether conda/virtual environment for text required python pacakges exists
}
\keyword{internal}
