% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleDataDescription.R
\docType{data}
\name{Language_based_assessment_data_3_100}
\alias{Language_based_assessment_data_3_100}
\title{Example text and numeric data.}
\format{
A data frame with 100 rows and 4 variables:
\describe{
   \item{harmonywords}{Word responses from the harmony in life word question}
  \item{hilstotal}{total score of the Harmony In Life Scale}
  \item{swlstotal}{total score of the Satisfaction With Life Scale}
}
}
\usage{
Language_based_assessment_data_3_100
}
\description{
The dataset is a shortened version of the data sets of Study 3-5
from Kjell, Kjell, Garcia and Sikström 2018.
}
\keyword{datasets}
