% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleDataDescription.R
\docType{data}
\name{Language_based_assessment_data_8}
\alias{Language_based_assessment_data_8}
\title{Text and numeric data for 10 participants.}
\format{
A data frame with 40 participants and 8 variables:
\describe{
  \item{harmonywords}{descriptive words where respondents describe their harmony in life}
  \item{satisfactionwords}{descriptive words where respondents describe their satisfaction with life}
  \item{harmonytexts}{text where respondents describe their harmony in life}
  \item{satisfactiontexts}{text where respondents describe their satisfaction with life}
  \item{hilstotal}{total score of the Harmony In Life Scale}
  \item{swlstotal}{total score of the Satisfaction With Life Scale}
  \item{age}{respondents age in years}
  \item{gender}{respondents gender 1=male, 2=female}
}
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/37126041/}
}
\usage{
Language_based_assessment_data_8
}
\description{
The dataset is a shortened version of the data sets of Study 3-5
from Kjell et al., (2018; https://psyarxiv.com/er6t7/).
}
\keyword{datasets}
