% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleDataDescription.R
\docType{data}
\name{PC_projections_satisfactionwords_40}
\alias{PC_projections_satisfactionwords_40}
\title{Example data for plotting a Principle Component Projection Plot.}
\format{
A data frame.
\describe{
  \item{words}{unique words}
  \item{n}{overall word frequency}
  \item{Dim_PC1}{Principle component value for dimension 1}
  \item{Dim_PC2}{Principle component value for dimension 2}
}
}
\source{
\url{https://link.springer.com/article/10.1007/s11205-015-0903-z}
}
\usage{
PC_projections_satisfactionwords_40
}
\description{
The dataset is a shortened version of the data sets of Study 1
from Kjell, et al., 2016.
}
\keyword{datasets}
