% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleDataDescription.R
\docType{data}
\name{raw_embeddings_1}
\alias{raw_embeddings_1}
\title{Word embeddings from textEmbedRawLayers function}
\format{
A list with token-level word embeddings for harmony words.
\describe{
  \item{tokens}{words}
  \item{layer_number }{layer of the transformer model}
  \item{Dim1:Dim8}{Word embeddings dimensions}
}
}
\source{
\url{https://osf.io/preprints/psyarxiv/er6t7/}
}
\usage{
raw_embeddings_1
}
\description{
The dataset is a shortened version of the data sets of Study 3-5
from Kjell, Kjell, Garcia and Sikström 2018.
}
\keyword{datasets}
