% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_5_textTrainPredictExamples.R
\name{reorder_columns}
\alias{reorder_columns}
\title{Reorder columns in a given dataframe}
\usage{
reorder_columns(df, target_col, desired_position)
}
\arguments{
\item{df}{(dataframe) Dataframe including the varibles that should get jitter.}

\item{target_col}{(string) name of the column that should have specified position.}

\item{desired_position}{(numeric) Position of the specified column.}
}
\description{
Target column and desired position
from the topicsScatterLegend() function.
}
