% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleDataDescription.R
\docType{data}
\name{sq_data_plottingHw_HILSSSWLS_100}
\alias{sq_data_plottingHw_HILSSSWLS_100}
\title{Example data for plotting a Dot Product Projection Plot.}
\format{
A data frame with 583 rows and 12 variables:
\describe{
  \item{words}{output from textProjectionData}
  \item{dot.x}{output from textProjectionData}
  \item{p_values_dot.x}{output from textProjectionData}
  \item{n_g1.x}{output from textProjectionData}
  \item{n_g2.x}{output from textProjectionData}
  \item{dot.y}{output from textProjectionData}
  \item{p_values_dot.y}{output from textProjectionData}
  \item{n_g1.y}{output from textProjectionData}
  \item{n_g2.y}{output from textProjectionData}
  \item{n}{output from textProjectionData}
  \item{n.percent}{output from textProjectionData}
  \item{N_participant_responses}{output from textProjectionData}
  ...
}
}
\source{
\url{https://psyarxiv.com/er6t7/}
}
\usage{
sq_data_plottingHw_HILSSSWLS_100
}
\description{
The dataset is a shortened version of the data sets of Study 3-5
from Kjell, Kjell, Garcia and Sikström 2018.
}
\keyword{datasets}
