% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleDataDescription.R
\docType{data}
\name{sq_data_tutorial4_100}
\alias{sq_data_tutorial4_100}
\title{Example text and numeric data.}
\format{A data frame with 608 rows and 4 variables:
\describe{
  \item{hilstotal}{total score of the Harmony In Life Scale}
  \item{swlstotal}{total score of the Satisfaction With Life Scale}
  \item{age}{respondents age in years}
  \item{gender}{respondents gender 1=male, 2=female}
  ...
}}
\source{
\url{https://psyarxiv.com/er6t7/}
}
\usage{
sq_data_tutorial4_100
}
\description{
The dataset is a shortened version of the data sets of Study 3-5
from Kjell, Kjell, Garcia and Sikström 2018.
}
\keyword{datasets}
