% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleDataDescription.R
\docType{data}
\name{sq_data_tutorial_plotting_hilswl}
\alias{sq_data_tutorial_plotting_hilswl}
\title{Example data for plotting.}
\format{
A data frame with 43 rows and 13 variables:
\describe{
  \item{words}{output from textPlotData}
  \item{n.x}{output from textPlotData}
  \item{dot.x}{output from textPlotData}
  \item{p_values_dot.x}{output from textPlotData}
  \item{n_g1.x}{output from textPlotData}
  \item{n_g2.x}{output from textPlotData}
   \item{n_all.x}{output from textPlotData}
  \item{Within_cosine_similarity.x}{output from textPlotData}
  \item{p_values_cosine.x}{output from textPlotData}
  \item{n.y}{output from textPlotData}
  \item{dot.y}{output from textPlotData}
  \item{p_values_dot.y}{output from textPlotData}
  \item{n_g1.y}{output from textPlotData}
  \item{n_g2.y}{output from textPlotData}
   \item{n_all.y}{output from textPlotData}
  \item{Within_cosine_similarity.y}{output from textPlotData}
  \item{p_values_cosine.y}{output from textPlotData}
  ...
}
}
\source{
\url{https://psyarxiv.com/er6t7/}
}
\usage{
sq_data_tutorial_plotting_hilswl
}
\description{
The dataset is a shortened version of the data sets of Study 3-5
from Kjell, Kjell, Garcia and Sikström 2018.
}
\keyword{datasets}
