% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_1_textPlot.R
\name{textCentralityData}
\alias{textCentralityData}
\title{textCentralityData computes cosine semantic similairty score between single words and the aggreagated word embedding of all words.}
\usage{
textCentralityData(
  words,
  wordembeddings,
  single_wordembeddings = single_wordembeddings_df,
  aggregation = "mean",
  min_freq_words = 0
)
}
\arguments{
\item{words}{Words or text variable to be plotted.}

\item{wordembeddings}{Wordembeddings from textEmbedd for the words to be plotted (i.e., the aggregated word embeddings for the "words" variable).}

\item{single_wordembeddings}{Wordembeddings from textEmbedd for individual words.}

\item{aggregation}{Method to aggreate the word embeddings (default = "mean"; see also, "min", "max", or "normalise1").}

\item{min_freq_words}{Option to only select words that have occured a specified number of times (default = 0); when creating the dot product projection
line (although single words recieve dot product projection and p-value) and within cosine similairty.}
}
\value{
A dataframe with variables (e.g., including semantic simikarity, frequencies) for the individual words
that is used for the plotting in the textCentralityPlot function.
}
\description{
textCentralityData computes cosine semantic similairty score between single words and the aggreagated word embedding of all words.
}
\examples{
wordembeddings <- wordembeddings4_10
data <- sq_data_tutorial8_10
df_for_plotting <- textCentralityData(data$harmonywords,
                                      wordembeddings$harmonywords,
                                      wordembeddings$singlewords_we
)
df_for_plotting
}
\seealso{
see \code{\link{textCentralityPlot}}  \code{\link{textProjectionData}}
}
