% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_3_textClassify.R
\name{textClassify}
\alias{textClassify}
\title{Predict label and probability of a text using a pretrained classifier language model.}
\usage{
textClassify(
  x,
  model,
  device = "cpu",
  tokenizer_parallelism = FALSE,
  logging_level = "error"
)
}
\arguments{
\item{x}{A character variable or a tibble/dataframe with at least one character variable.}

\item{model}{Character string specifying a pre-trained classifier language model.
For full list of options see pretrained classifier models at
\href{https://huggingface.co/transformers/pretrained_models.html}{HuggingFace}.
For example use "cardiffnlp/twitter-roberta-base-sentiment", "distilbert-base-uncased-finetuned-sst-2-english".}

\item{device}{Name of device to use: 'cpu', 'gpu', or 'gpu:k' where k is a specific device number}

\item{tokenizer_parallelism}{If TRUE this will turn on tokenizer parallelism. Default FALSE.}

\item{logging_level}{Set the logging level. Default: "warning".
Options (ordered from less logging to more logging): critical, error, warning, info, debug}
}
\value{
A tibble with predicted labels and scores for each text variable.
The comment of the object show the model-name and computation time.
}
\description{
Predict label and probability of a text using a pretrained classifier language model.
}
\examples{
\donttest{
classifications <- textClassify(x = Language_based_assessment_data_8[1:2, 1:2])
classifications
comment(classifications)
}
}
\seealso{
see \code{\link{textEmbedLayerAggregation}} and \code{\link{textEmbed}}
}
