% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_0_textDescriptives.R
\name{textClean}
\alias{textClean}
\title{Cleans text from standard personal information}
\usage{
textClean(
  text,
  replace = TRUE,
  date = TRUE,
  time = TRUE,
  phone = TRUE,
  email = TRUE,
  ip = TRUE,
  money = TRUE,
  creditcard = TRUE,
  bitcoin = TRUE,
  location = TRUE,
  ssn = TRUE,
  at_symbol = TRUE,
  url = TRUE
)
}
\arguments{
\item{text}{(character) The text to be cleaned.}

\item{replace}{(boolean)}

\item{date}{(boolean)}

\item{time}{(boolean)}

\item{phone}{(boolean)}

\item{email}{(boolean)}

\item{ip}{(boolean)}

\item{money}{(boolean)}

\item{creditcard}{(boolean)}

\item{bitcoin}{(boolean)}

\item{location}{(boolean)}

\item{ssn}{(boolean)}

\item{at_symbol}{(boolean)}

\item{url}{(boolean)}
}
\value{
Text cleaned from typical personal identifiable information
}
\description{
The text is being cleaned from information that may identify them; however, note that
this is not a guarantee for anonymization.
}
