% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_1_textEmbed.R
\name{textCleanNonASCII}
\alias{textCleanNonASCII}
\title{Clean non-ASCII characters}
\usage{
textCleanNonASCII(data_tibble)
}
\arguments{
\item{data_tibble}{A tibble with character variables.}
}
\value{
a tibble with removed ascii characters
}
\description{
textCleanNonASCII() cleans all text entries with a non-ASCII character in a tibble.
}
