% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_1_textEmbed.R
\name{textCleanNonASCIIinfo}
\alias{textCleanNonASCIIinfo}
\title{Clean non-ASCII characters}
\usage{
textCleanNonASCIIinfo(data_tibble, problematic_texts = NULL, remove_non_ascii)
}
\arguments{
\item{data_tibble}{A tibble with character variables.}

\item{problematic_texts}{(tibble)}

\item{remove_non_ascii}{(boolean)}
}
\value{
a tibble with removed ascii characters
}
\description{
textCleanNonASCII() cleans all text entries with a non-ASCII character in a tibble.
}
