% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_0_2_textDiagnostics.R
\name{textDiagnostics}
\alias{textDiagnostics}
\title{Print diagnostic information for the text package environment}
\usage{
textDiagnostics(
  anonymise = FALSE,
  search_other_envs = FALSE,
  search_omp = TRUE
)
}
\arguments{
\item{anonymise}{logical; if TRUE, personal paths and usernames are redacted before displaying output (default is FALSE).}

\item{search_other_envs}{search different different python environments}

\item{search_for_omp}{search for OMP related packages (most relevant for MacOS, that can get conflicts).}
}
\description{
Print diagnostic information for the text package environment
}
