% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_2_textDiff.R
\name{textDiff}
\alias{textDiff}
\title{textDiff: Test whether there is a significant difference between two sets of texts
(i.e., between their word embeddings). This is achieved using permutation, in the following steps:}
\usage{
textDiff(
  x,
  y,
  Npermutations = 1000,
  method = "paired",
  alternative = c("two_sided", "less", "greater"),
  output.permutations = TRUE,
  N_cluster_nodes = 1
)
}
\arguments{
\item{x}{set of word embeddings from textEmbedd.}

\item{y}{set of word embeddings from textEmbedd.}

\item{Npermutations}{number of permatation (default 1000).}

\item{method}{compute a "paired" or "unpaired" test.}

\item{alternative}{wheather to use a two or one-sided test.}

\item{output.permutations}{If TRUE returns permuted values in output.}

\item{N_cluster_nodes}{Number of cluster nodes to use (more makes computation faster; see parallel package.}
}
\value{
A list with a p-value, estimate and perumted values if output.permutations=TRUE.
}
\description{
textDiff: Test whether there is a significant difference between two sets of texts
(i.e., between their word embeddings). This is achieved using permutation, in the following steps:
}
\examples{
x <- wordembeddings4_10$harmonywords
y <- wordembeddings4_10$satisfactionwords
textDiff(x, y, method = "paired", Npermutations = 10, N_cluster_nodes = 1)
}
