% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_1_textSimilarity.R
\name{textDistanceMatrix}
\alias{textDistanceMatrix}
\title{Compute semantic distance scores between all combinations in a word embedding}
\usage{
textDistanceMatrix(x, method = "euclidean")
}
\arguments{
\item{x}{Word embeddings (from textEmbed).}

\item{method}{Character string describing type of measure to be computed; default is "euclidean" (see also
measures from stats:dist() including "maximum", "manhattan", "canberra", "binary" and "minkowski".
It is also possible to use "cosine", which computes the cosine distance (i.e., 1 - cosine(x, y)).}
}
\value{
A matrix of semantic distance scores
}
\description{
Compute semantic distance scores between all combinations in a word embedding
}
\examples{
distance_scores <- textDistanceMatrix(word_embeddings_4$texts$harmonytext[1:3, ])
round(distance_scores, 3)
}
\seealso{
see \code{\link{textDistanceNorm}} and \code{\link{textSimilarityTest}}
}
