% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_5_textDomainCompare.R
\name{textDomainCompare}
\alias{textDomainCompare}
\title{Compare two language domains}
\usage{
textDomainCompare(train_language, assess_language)
}
\arguments{
\item{train_language}{A word-frequency data frame from textTokenizeAndCount}

\item{assess_language}{A word-frequency data frame from textTokenizeAndCount}
}
\value{
List with similarity scores: overlapp_percentage, test_recall_percentage and cosine_similarity
}
\description{
Compare two language domains
}
\examples{
\dontrun{
train_language <- textTokenizeAndCount(Language_based_assessment_data_8["harmonytexts"])
assess_language <- textTokenizeAndCount(Language_based_assessment_data_8["satisfactiontexts"])
textDomainCompare(train_language, assess_language)
}
}
\seealso{
see \code{\link{textTokenizeAndCount}}
}
