% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_1_textEmbed.R
\name{textEmbed}
\alias{textEmbed}
\title{textEmbed extracts layers and aggregate them to word embeddings, for all character variables in the dataframe.}
\usage{
textEmbed(
  x,
  contexts = TRUE,
  decontexts = TRUE,
  pretrained_weights = "bert-base-uncased",
  tokenizer_class = BertTokenizer,
  model_class = BertModel,
  layers = 11,
  context_aggregation = "mean",
  context_tokens_select = NULL,
  context_tokens_deselect = NULL,
  decontext_layers = 11:12,
  decontext_return_tokens = TRUE,
  decontext_aggregation = "mean",
  decontext_tokens_select = NULL,
  decontext_tokens_deselect = NULL
)
}
\arguments{
\item{x}{Tibble/dataframe with at least one character variables.}

\item{contexts}{provides word embeddings based on contexts (default = TRUE).}

\item{decontexts}{provides word embeddings of single words as input (default = TRUE), which are used for plotting.}

\item{pretrained_weights}{Character specifying pre-trained language model through RBERT. Current options
"bert_base_uncased", "bert_base_cased", "bert_large_uncased",
"bert_large_cased", "bert_large_uncased_wwm", "bert_large_cased_wwm",
"bert_base_multilingual_cased", "bert_base_chinese", "scibert_scivocab_uncased",
"scibert_scivocab_cased", "scibert_basevocab_uncased", and "scibert_basevocab_cased".}

\item{tokenizer_class}{tokenizer that match pretrained_weights}

\item{model_class}{model class that matches pretrained_weights and tokenizer_class}

\item{layers}{most efficient to only extract the ones you need; but can also extract 'all',
and then remove in textLayerAggregation function, when aggregating embeddings.}

\item{context_aggregation}{method to aggregate the contextualised layers (e.g., "mean", "min" or "max).}

\item{context_tokens_select}{option to select embeddings linked to specifc tokens such as [CLS] and [SEP] for the context embeddings.}

\item{context_tokens_deselect}{option to select embeddings linked to specifc tokens such as [CLS] and [SEP] for the decontext embeddings.}

\item{decontext_layers}{layers to use for the decontext embeddings.}

\item{decontext_return_tokens}{TRUE or FALSE for returning tokens.}

\item{decontext_aggregation}{Method to aggregate the decontextualised layers (i.e., embeddings from single words; e.g., "mean", "min" or "max).}

\item{decontext_tokens_select}{option to select embeddings linked to specifc tokens such as [CLS] and [SEP] for the decontext embeddings.}

\item{decontext_tokens_deselect}{option to deselect embeddings linked to specifc tokens such as [CLS] and [SEP] for the decontext embeddings.}
}
\value{
A tibble with tokens, layer identifyer and word embeddings. Note that layer 0 is the input embedding to the transformer
}
\description{
textEmbed extracts layers and aggregate them to word embeddings, for all character variables in the dataframe.
}
\seealso{
see \code{\link{textLayerAggregation}} and \code{\link{textEmbed}}
}
