% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_1_textEmbed.R
\name{textFindNonASCII}
\alias{textFindNonASCII}
\title{# Function to detect non-ASCII characters in a tibble with multiple columns}
\usage{
textFindNonASCII(data_tibble)
}
\arguments{
\item{data_tibble}{A character variable or a tibble including  character variables.}
}
\value{
a tibble containing variable names, row numbers and text including non-acii.
}
\description{
# Function to detect non-ASCII characters in a tibble with multiple columns
}
