% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_2_textGeneration.R
\name{textGeneration}
\alias{textGeneration}
\title{Predicts the words that will follow a specified text prompt. STILL UNDER DEVELOPMENT}
\usage{
textGeneration(
  x,
  model = "gpt2",
  device = "cpu",
  tokenizer_parallelism = FALSE,
  logging_level = "warning",
  return_incorrect_results = FALSE,
  return_tensors = FALSE,
  return_text = TRUE,
  return_full_text = TRUE,
  clean_up_tokenization_spaces = FALSE,
  prefix = ""
)
}
\arguments{
\item{x}{(string)  A character variable or a tibble/dataframe with at least one character variable.}

\item{model}{(string)  Specify a pre-trained language model that have been trained with an autoregressive language modeling
objective, which includes the uni-directional models (e.g., gpt2).}

\item{device}{(string)  Name of device to use: 'cpu', 'gpu', or 'gpu:k' where k is a specific device number}

\item{tokenizer_parallelism}{(boolean)  If TRUE this will turn on tokenizer parallelism. Default FALSE.}

\item{logging_level}{(string)  Set the logging level. Default: "warning".
Options (ordered from less logging to more logging): critical, error, warning, info, debug}

\item{return_incorrect_results}{(boolean)  Many models are not created to be able to provide classifications - this setting
stops them from returning incorrect results.}

\item{return_tensors}{(boolean)  Whether or not the output should include the prediction tensors (as token indices).}

\item{return_text}{(boolean)  Whether or not the outputs should include the decoded text.}

\item{return_full_text}{(boolean) If FALSE only the added text is returned, otherwise the full text is returned.
(This setting is only meaningful if return_text is set to TRUE)}

\item{clean_up_tokenization_spaces}{(boolean)  Option to clean up the potential extra spaces in the returned text.}

\item{prefix}{(string) Option to add a prefix to prompt.}
}
\value{
A tibble with generated text.
}
\description{
Predicts the words that will follow a specified text prompt. STILL UNDER DEVELOPMENT
}
\examples{
\donttest{
# generated_text <- textGeneration("The meaning of life is")
# generated_text
}
}
\seealso{
see \code{\link{textEmbedLayerAggregation}} and \code{\link{textEmbed}}
}
