% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_99_textImport.R
\name{textImport}
\alias{textImport}
\title{textImport extracts word embeddings for all text in character variables in dataframe}
\usage{
textImport(
  x,
  model = "bert_base_uncased",
  layer_indexes_RBERT = 12,
  batch_size = 2L,
  token_index_filter = 1,
  layer_index_filter = 12,
  ...
)
}
\arguments{
\item{x}{Tibble.}

\item{model}{Character specifying pre-trained language model through RBERT. Current options
"bert_base_uncased", "bert_base_cased", "bert_large_uncased",
"bert_large_cased", "bert_large_uncased_wwm", "bert_large_cased_wwm",
"bert_base_multilingual_cased", "bert_base_chinese", "scibert_scivocab_uncased",
"scibert_scivocab_cased", "scibert_basevocab_uncased", and "scibert_basevocab_cased".}

\item{layer_indexes_RBERT}{layers to be used from RBERT (default 12; in REBERT it is 1:12).}

\item{batch_size}{batch size from RBERT (default 2L)}

\item{token_index_filter}{from RBERT}

\item{layer_index_filter}{from RBERT}

\item{...}{arguments from RBERT function extract_features}
}
\value{
A list with word embeddings.
}
\description{
textImport extracts word embeddings for all text in character variables in dataframe
}
\examples{
x <- sq_data_tutorial8_10[1:2, 1:2]
library(dplyr)
wordembeddings <- textImport(x)
}
\seealso{
see \code{\link{textTrain}}, \code{\link{textTtest}} and \code{\link{textSimilarity}}
}
