% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_textImport.R
\name{textImportWords}
\alias{textImportWords}
\title{textImportWords extracts word embeddings on single words basis and adds them together
for all character variables in given tibble}
\usage{
textImportWords(
  x,
  layer_indexes_RBERT = 12,
  batch_size_IBT = 2L,
  token_index_IBT = 1,
  layer_index_IBT = 12,
  Ndim = 768,
  ...
)
}
\arguments{
\item{x}{Tibble.}

\item{layer_indexes_RBERT}{layers to be used from RBERT (default 12; in REBERT it is 1:12).}

\item{batch_size_IBT}{batch size from RBERT (default 2L)}

\item{token_index_IBT}{from RBERT}

\item{layer_index_IBT}{from RBERT}

\item{Ndim}{Number of dimensions in the word embedding}

\item{...}{arguments from RBERT function extract_features}
}
\value{
A list with word embeddings.
}
\description{
textImportWords extracts word embeddings on single words basis and adds them together
for all character variables in given tibble
}
\examples{
x <- sq_data_tutorial8_10[1:2, 1:2]
library(dplyr)
wordembeddings <- textImportWords(x)
}
\seealso{
see \code{\link{textImport}}, \code{\link{textTtest}} and \code{\link{textSimilarity}}
}
