% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_1_textEmbed.R
\name{textLayerAggregation}
\alias{textLayerAggregation}
\title{textLayerAggregation selects and aggregate layers of hidden states to get word embeddings.}
\usage{
textLayerAggregation(
  word_embeddings_layers,
  layers = "all",
  aggregation = "mean",
  tokens_select = NULL,
  tokens_deselect = NULL
)
}
\arguments{
\item{word_embeddings_layers}{Layers outputted from textHuggingFace.}

\item{layers}{state the the numbers of the layers that you want to select for aggregation (e.g., c(1, 10:11) to get the first, tenth and eleventh).
Note that layer 0 is the input embedding to the transformer.}

\item{aggregation}{aggregation method of the dimension of each layer selected with layers parameter above (e.g., "mean", "min" or "max).}

\item{tokens_select}{option to select embeddings linked to specifc tokens such as [CLS] and [SEP].}

\item{tokens_deselect}{option to deselect embeddings linked to specifc tokens such as [CLS] and [SEP].}
}
\value{
A tibble with word embeddings. Note that layer 0 is the input embedding to the transformer.
}
\description{
textLayerAggregation selects and aggregate layers of hidden states to get word embeddings.
}
\seealso{
see \code{\link{textHuggingFace}} and \code{\link{textEmbed}}
}
