% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_4_0_textPredict_Assess_Classify.R
\name{textLBAM}
\alias{textLBAM}
\title{The LBAM library}
\usage{
textLBAM(columns = NULL, lbam_update = FALSE)
}
\arguments{
\item{columns}{(string) Select which columns to retrieve e.g., c("Name", "Path")}

\item{lbam_update}{(boolean) TRUE downloads a new copy of the LBAM file}
}
\value{
Data frame containing information about the Language-based assessment models library (LBAM).
}
\description{
Retrieve the Language-Based Assessment Models library (LBAM).
}
\examples{
\dontrun{
library(dplyr)
test_lbam <- textLBAM(
  lbam_update = TRUE
)

test_lbam \%>\%
  filter(substr(Construct_Concept_Behaviours, 1, 3) == "Dep") \%>\%
  dplyr::select(Name)
}
}
