% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_3_textModels.R
\name{textModelLayers}
\alias{textModelLayers}
\title{Get the number of layers in a given model.}
\usage{
textModelLayers(target_model)
}
\arguments{
\item{target_model}{(string) The name of the model to know the number of layers of.}
}
\value{
Number of layers.
}
\description{
Get the number of layers in a given model.
}
\examples{
\dontrun{
textModelLayers(target_model = "bert-base-uncased")
}
}
\seealso{
see \code{\link{textModels}}
}
