% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_1_textEmbed.R
\name{textModels}
\alias{textModels}
\title{Check models.}
\usage{
textModels()
}
\value{
List of names of models on your machine
}
\description{
Check models.
}
\examples{
\donttest{
textModel()
}
}
\seealso{
see \code{\link{textModelsRM}}
}
