% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_1_textEmbed.R
\name{textModelsRM}
\alias{textModelsRM}
\title{Delete huggingface models}
\usage{
textModelsRM(target_model)
}
\description{
Delete huggingface models
}
\examples{
\donttest{
textModelsRM("name-of-model")
}
}
\seealso{
see \code{\link{textModels}}
}
