% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_2_textPlotPCA.R
\name{textPCA}
\alias{textPCA}
\title{Compute 2 PCA dimensions of the word embeddings for individual words.}
\usage{
textPCA(words, single_word_embeddings = single_word_embeddings_df, seed = 1010)
}
\arguments{
\item{words}{Word or text variable to be plotted.}

\item{single_word_embeddings}{Word embeddings from textEmbed for individual words
(i.e., decontextualized embeddings).}

\item{seed}{Set different seed.}
}
\value{
A dataframe with words, their frquency and two PCA dimensions from the word_embeddings
for the individual words that is used for the plotting in the textPCAPlot function.
}
\description{
Compute 2 PCA dimensions of the word embeddings for individual words.
}
\examples{
# Data
df_for_plotting2d <- textPCA(
  words = Language_based_assessment_data_8$harmonywords,
  single_word_embeddings = word_embeddings_4$singlewords_we
)
df_for_plotting2d
}
\seealso{
see \code{\link{textPCAPlot}}
}
