% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_1_textPlot.R
\name{textPlotData}
\alias{textPlotData}
\title{textPlotData computes variables for plotting words.}
\usage{
textPlotData(
  words,
  wordembeddings,
  single_wordembeddings = single_wordembeddings_df,
  x,
  y = NULL,
  pca = NULL,
  aggregation = "mean",
  split = "quartile",
  word_weight_power = 1,
  min_freq_words = 0,
  Npermutations = 10000,
  n_per_split = 50000
)
}
\arguments{
\item{words}{Words or text variable to be plotted.}

\item{wordembeddings}{Wordembeddings from textEmbedd for the words to be plotted (i.e., the aggregated word embeddings for the "words" variable).}

\item{single_wordembeddings}{Wordembeddings from textEmbedd for individual words.}

\item{x}{Numeric variable the words should be plotted according to.}

\item{y}{Numeric variable the words should be plotted according to (y=NULL).}

\item{pca}{Number of pca dimensions applied to the word embeddings in the beginning of the function. A number below 1 takes out \% of variance;
An integer specify number of components to extract.}

\item{aggregation}{Method to aggreate the word embeddings (default = "mean"; see also, "min", "max", or "normalise1").}

\item{split}{Method to split the axes (default = "quartile" (i.e., selecting lower and upper quartile); see also "median").}

\item{word_weight_power}{Takes the power of the frequency in the computation of aggreated word embeddings for group low (1) and high (2).}

\item{min_freq_words}{Option to only select words that have occured a specified number of times (default = 0); when creating the dot product projection
line (although single words recieve dot product projection and p-value) and within cosine similairty.}

\item{Npermutations}{Number of permutations in the creation of the null distribution.}

\item{n_per_split}{A setting to split Npermutations to avoid reaching computer memory limits; the higher the faster, but too high may lead to abortion.}
}
\value{
A dataframe with variables (e.g., including dot product, frequencies, p-values) for the individual words
that is used for the plotting in the textPlotViz function.
}
\description{
textPlotData computes variables for plotting words.
}
\examples{
wordembeddings <- wordembeddings4_10
data <- sq_data_tutorial8_10
df_for_plotting <- textPlotData(data$harmonywords, wordembeddings$harmonywords,
  wordembeddings$singlewords_we,
  data$hilstotal,
  split = "quartile",
  Npermutations = 10,
  n_per_split = 1
)
df_for_plotting
}
\seealso{
see \code{\link{textPlotViz}}
}
