% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_textPlot.R
\name{textPlotViz}
\alias{textPlotViz}
\title{textPlotViz trains word embeddings to a numeric variable.}
\usage{
textPlotViz(
  word_data,
  top_p_words = -25,
  x_axes = "cohensD.x",
  y_axes = "cohensD.y",
  x_axes_label = "Cohen's D",
  y_axes_label = "Cohen's D",
  scale_x_axes_lim = NULL,
  scale_y_axes_lim = NULL,
  y_axes_values = NULL,
  word_font = "Arial",
  colors_words = c("#ff0000", "#ff8080", "white", "#1aff1a", "#00cc00"),
  colors_words_scale = c(-0.1, -0.01, 0, 0.01, 0.1),
  word_size_range = c(3, 8),
  position_jitter_hight = 0,
  position_jitter_width = 0.03,
  point_size = 0.5,
  arrow_transparency = 0.1
)
}
\arguments{
\item{word_data}{Dataframe from textPlotData.}

\item{top_p_words}{Number of significant words to plot (default -25, which plots the 25 most significant words).}

\item{x_axes}{Variable to be plotted on the x axes (default is "cohensD.x")}

\item{y_axes}{Variable to be plotted on the y axes (default is "cohensD.y"). To only print 1-dimension insert NULL.}

\item{x_axes_label}{"Cohen's D",}

\item{y_axes_label}{"p-value",}

\item{scale_x_axes_lim}{NULL,}

\item{scale_y_axes_lim}{NULL,}

\item{y_axes_values}{NULL,}

\item{word_font}{"Arial",}

\item{colors_words}{c("#ff0000", "#ff8080", "white", "#1aff1a", "#00cc00"),}

\item{colors_words_scale}{c(-0.1, -0.01, 0, 0.01, 0.1),}

\item{word_size_range}{c(3, 8),}

\item{position_jitter_hight}{.0,}

\item{position_jitter_width}{.03,}

\item{point_size}{0.5,}

\item{arrow_transparency}{0.1}
}
\value{
A 1- or 2-dimensional word plot.
}
\description{
textPlotViz trains word embeddings to a numeric variable.
}
\examples{
word_plot_data
plot <- textPlotViz(word_data = word_plot_data,
     top_p_words = -25,
     x_axes = "cohensD.x",
     y_axes = "cohensD.y",
     x_axes_label = "HILS: Cohen's D",
     y_axes_label = "SWLS: Cohen's D",
     scale_x_axes_lim = NULL,
     scale_y_axes_lim = NULL,
     y_axes_values = NULL,
     word_font = "Arial",
     colors_words = c("#ff0000", "#ff8080", "white", "#1aff1a", "#00cc00"),
     colors_words_scale = c(-0.1, -0.01, 0, 0.01, 0.1),
     word_size_range = c(3, 8),
     position_jitter_hight = .0,
     position_jitter_width = .03,
     point_size = 0.5,
     arrow_transparency = 0.1)
plot
}
\seealso{
see \code{\link{textPlotData}}
}
