% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textPredictEntireProcedure.R
\name{textPredictEntireProcedure}
\alias{textPredictEntireProcedure}
\title{(experimental) An end-to-end, user-friendly function that takes text as input and provides predictions based on a pre-trained model.}
\usage{
textPredictEntireProcedure(
  text_to_predict = NULL,
  premade_embeddings = NULL,
  model_platform = "github",
  model_reference =
    "https://github.com/CarlViggo/pretrained-models/raw/main/trained_hils_model.RDS",
  save_model = FALSE,
  model_name = NULL,
  type = "class",
  max_token_to_sentence = 4,
  aggregation_from_layers_to_tokens = "concatenate",
  aggregation_from_tokens_to_texts = "mean"
)
}
\arguments{
\item{text_to_predict}{(character) Text to predict. If this argument is specified, then argument "premade_embeddings" must be set to NULL (default = NULL).}

\item{premade_embeddings}{(Embeddings from e.g., textEmbed) Embeddings to predict. If this argument is specified, then argument "texts" must be set to NULL (default = NULL).}

\item{model_platform}{(character) Either "github" or "local" (default = "github").}

\item{model_reference}{(character) Link to github-model.}

\item{save_model}{(boolean) If set to true, the model will be saved in work-directory (default = FALSE). If TRUE, then argument "model_name" must be defined.}

\item{model_name}{(character) Optional. If the 'save_model' argument is set to TRUE and you wish to assign a custom name to your model, 
provide a file path along with the desired filename, for example, 'C:/Users/Name/Models/linear_model.RDS' (default is NULL). If this is not defined 
but 'save_model' is set to TRUE, the model will be saved as 'imported_model.RDS'.}

\item{type}{(character) Choose either 'class' or 'prob'. If your model is a logistic or multinomial model, specify whether you want to receive the 
model's classification "class" or the underlying probabilities "prob" (default = "class").}

\item{max_token_to_sentence}{(numeric) This information will be automatically extracted from your model, so this argument is typically not used. 
However, if it's not defined in your model_description, you can assign a value here (default = 4).}

\item{aggregation_from_layers_to_tokens}{(character) This information will be automatically extracted from your model, so this argument is typically not used. 
However, if it's not defined in your model_description, you can assign a value here (default = "concatenate").}

\item{aggregation_from_tokens_to_texts}{(character) This information will be automatically extracted from your model, so this argument is typically not used. 
However, if it's not defined in your model_description, you can assign a value here (default = "mean").}
}
\value{
The textPredictEntireProcedure function takes text as input and returns predictions. The type of prediction returned depends 
on the specific pre-trained model.
}
\description{
(experimental) An end-to-end, user-friendly function that takes text as input and provides predictions based on a pre-trained model.
}
\examples{
# (data from Language_based_assessment_data_8)
\dontrun{
  input <- "I am not in harmony in my life as much as I would like to be.  
    I would like to be more peaceful and more productive.  I want to have a balanced life.  
    I would like to spend more time with my family and be  more focused on my goals and where 
    I want to be in life and what I want to achieve."
 
# Example 1 (Predict valence from input): 
  prediction <- textPredictEntireProcedure(text_to_predict = input)
  
# Example 2 (Predict valence from premade word-embeddings) (data from Language_based_assessment_data_8): 
  
  # create embeddings
  embeddings <- textEmbed("I am not in harmony in my life as much as I would like to be.")
  
  # predict from embeddings
  prediction <- textPredictEntireProcedure(premade_embeddings = embeddings)

# Example 3 (Use a logistic model to predict gender, return the underlying probabilities and save model): 
  prediction <- textPredictEntireProcedure(text_to_predict = input, 
                             model_reference = "https://github.com/CarlViggo/pretrained_swls_model/raw/main/trained_github_model_logistic.RDS", 
                             save_model = TRUE, 
                             type = "prob")
}

}
\seealso{
See \code{\link{textPredict}} and \code{\link{textpredictAll}}.
}
