% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_1_textPlot.R
\name{textProjectionPlot}
\alias{textProjectionPlot}
\title{textProjectionPlot plots words}
\usage{
textProjectionPlot(
  word_data,
  k_n_words_two_test = TRUE,
  min_freq_words = 1,
  plot_n_words_square = 3,
  plot_n_words_p = 5,
  plot_n_word_extreme = 5,
  plot_n_word_frequency = 5,
  plot_n_words_middle = 5,
  titles_color = "#61605e",
  x_axes = "dot.x",
  y_axes = NULL,
  p_values_x = "p_values_dot.x",
  p_values_y = NULL,
  p_alpha = 0.05,
  p_adjust_method = "none",
  title_top = " Dot Product Projection",
  x_axes_label = "Dot product projection (DPP)",
  y_axes_label = "Dot product projection (DPP)",
  scale_x_axes_lim = NULL,
  scale_y_axes_lim = NULL,
  y_axes_values = element_blank(),
  word_font = NULL,
  bivariate_color_codes = c("#398CF9", "#60A1F7", "#5dc688", "#e07f6a", "#EAEAEA",
    "#40DD52", "#FF0000", "#EA7467", "#85DB8E"),
  word_size_range = c(3, 8),
  position_jitter_hight = 0,
  position_jitter_width = 0.03,
  point_size = 0.5,
  arrow_transparency = 0.1,
  points_without_words_size = 0.2,
  points_without_words_alpha = 0.2,
  legend_title = "DPP",
  legend_x_axes_label = "x",
  legend_y_axes_label = "y",
  legend_x_position = 0.02,
  legend_y_position = 0.02,
  legend_h_size = 0.2,
  legend_w_size = 0.2,
  legend_title_size = 7,
  legend_number_size = 2
)
}
\arguments{
\item{word_data}{Dataframe from textPlotData.}

\item{k_n_words_two_test}{Selects the k most frequent words to significance test (k = sqrt(100*N); N = number of paricipant responses). Default = TRUE.}

\item{min_freq_words}{Select words to signficance test that has occured at least min_freq_words (default = 1).}

\item{plot_n_words_square}{Select number of significant words in each square to plot.}

\item{plot_n_words_p}{Number of significant words to plot (n per x and y axes, where duplicates are removed);
selects fist according to lowest p-value and then to frequency (default 25  per axes, where duplicates are removed).}

\item{plot_n_word_extreme}{Number of words that are extreme on dot product projection per dimension.
(i.e., even if not significant;  per dimensions, where duplicates are removed).}

\item{plot_n_word_frequency}{Number of words based on being most frequent.
(i.e., even if not significant).}

\item{plot_n_words_middle}{Number of words plotted that are in the middle in dot product projection score
(i.e., even if not significant;  per dimensions, where duplicates are removed).}

\item{titles_color}{Color for all the titles (default: "#61605e")}

\item{x_axes}{Variable to be plotted on the x axes (default is "dot_product.x").}

\item{y_axes}{Variable to be plotted on the y axes (default is "cohensD.y"). To only print 1-dimension insert NULL.}

\item{p_values_x}{P-values variable to plott according to.}

\item{p_values_y}{P-values variable to plott according to.}

\item{p_alpha}{Alpha (default = .05).}

\item{p_adjust_method}{Method to adjust/correct p-value for multiple comparisons (deafult = "holm"; see also "none", "hochberg", "hommel", "bonferroni", "BH", "BY",  "fdr")}

\item{title_top}{string for title (default "  ")}

\item{x_axes_label}{Label on the x-axes.}

\item{y_axes_label}{Label on the y-axes.}

\item{scale_x_axes_lim}{Manually set the legth of the x-axes (deafult=NULL; e.g., try c(-5, 5)).}

\item{scale_y_axes_lim}{Manually set the legth of the y-axes (deafult=NULL; e.g., try c(-5, 5)).}

\item{y_axes_values}{NULL,}

\item{word_font}{Type of font (default: "Arial").}

\item{bivariate_color_codes}{The diffent colors of the words (default: c("#398CF9", "#60A1F7", "#5dc688",
"#e07f6a", "#EAEAEA", "#40DD52",
"#FF0000", "#EA7467", "#85DB8E")).}

\item{word_size_range}{a vector with minimum and maximum (default: c(3, 8)).}

\item{position_jitter_hight}{degree of jitter hight (default: .0).}

\item{position_jitter_width}{degree of jitter hight (default: .03).}

\item{point_size}{the size of the points indicating the words' position (default: 0.5).}

\item{arrow_transparency}{the transparency of the lines between each word and point (default: 0.1).}

\item{points_without_words_size}{sets the size of the points (without showing words, default is to not show it, i.e., 0).}

\item{points_without_words_alpha}{sets the transperency of the points (without showing words, default is to not show it, i.e., 0).}

\item{legend_title}{Label on the color legend (default: "(DPP)".}

\item{legend_x_axes_label}{Label on the color legend (default: "(x)".}

\item{legend_y_axes_label}{Label on the color legend (default: "(y)".}

\item{legend_x_position}{Position on the x coordinates of the color legend (default: 0.02).}

\item{legend_y_position}{Position on the y coordinates of the color legend (default: 0.05).}

\item{legend_h_size}{The height of the color legend (default 0.15).}

\item{legend_w_size}{The width of the color legend (default 0.15).}

\item{legend_title_size}{=7}

\item{legend_number_size}{size of the values in the legend (default = 2)}
}
\value{
A 1- or 2-dimensional word plot.
}
\description{
textProjectionPlot plots words
}
\examples{
# The test-data included in the package is called: sq_data_plottingHw_HILSSSWLS_100

# Dot Product Projection Plot
plot_projection <- textProjectionPlot(
 word_data = sq_data_plottingHw_HILSSSWLS_100,
 k_n_words_two_test = FALSE,
 min_freq_words = 1,
 plot_n_words_square = 3,
 plot_n_words_p = 3,
 plot_n_word_extreme = 1,
 plot_n_word_frequency = 1,
 plot_n_words_middle = 1,
 x_axes = "dot.x",
 y_axes = "dot.y",
 p_values_x = "p_values_dot.x",
 p_values_y = "p_values_dot.y",
 p_alpha = 0.05,
 title_top = " Dot Product Projection (DPP)",
 x_axes_label = "Low vs. High HILS score",
 y_axes_label = "Low vs. High SWLS score",
 p_adjust_method = "bonferroni",
 scale_y_axes_lim = NULL
)
plot_projection

names(sq_data_plottingHw_HILSSSWLS_100)

}
\seealso{
see \code{\link{textProjectionData}}
}
