% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_5_textQA.R
\name{textQA}
\alias{textQA}
\title{Question Answer STILL UNDER DEVELOPMENT}
\usage{
textQA(
  question,
  context,
  model = "",
  device = "cpu",
  tokenizer_parallelism = FALSE,
  logging_level = "warning",
  return_incorrect_results = FALSE,
  topk = 1,
  doc_stride = 128,
  max_answer_len = 15,
  max_seq_len = 384,
  max_question_len = 64,
  handle_impossible_answer = FALSE
)
}
\arguments{
\item{question}{(string)  A character variable or a tibble/dataframe with at least one character variable.}

\item{context}{(string)  The context(s) where the model will look for the answer.}

\item{model}{(string)  Specify a pre-trained language model that have been fine-tuned on a question answering task.}

\item{device}{(string)  Name of device to use: 'cpu', 'gpu', or 'gpu:k' where k is a specific device number}

\item{tokenizer_parallelism}{(boolean)  If TRUE this will turn on tokenizer parallelism. Default FALSE.}

\item{logging_level}{(string)  Set the logging level. Default: "warning".
Options (ordered from less logging to more logging): critical, error, warning, info, debug}

\item{return_incorrect_results}{(boolean)  Many models are not created to be able to provide classifications - this setting
stops them from returning incorrect results.}

\item{topk}{(integer) (int)  Indicates number of possible answer span(s) to get from the model output.}

\item{doc_stride}{(integer)   If the context is too long to fit with the question for the model, it will be split into overlapping chunks.
This setting controls the overlap size.}

\item{max_answer_len}{(integer)  Max answer size to be extracted from the model’s output.}

\item{max_seq_len}{(integer)  The max total sentence length (context + question) in tokens of each chunk passed to the model.
If needed, the context is split in chunks (using doc_stride as overlap).}

\item{max_question_len}{(integer)   The max question length after tokenization. It will be truncated if needed.}

\item{handle_impossible_answer}{(boolean)  Whether or not impossible is accepted as an answer.}
}
\value{
A tibble with.
}
\description{
Question Answer STILL UNDER DEVELOPMENT
}
\examples{
\donttest{
# qa_example <- textQA(question = "why are the ")
}
}
\seealso{
see \code{\link{textEmbedLayerAggregation}} and \code{\link{textEmbed}}
}
