% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_0_1_text_initialize.R
\name{textrpp_finalize}
\alias{textrpp_finalize}
\title{Finalize text required python packages}
\usage{
textrpp_finalize()
}
\description{
While running text required python packages on Python through R, a Python process is always running
in the background and the R session will take up a lot of memory (typically over
1.5GB). \code{textrpp_finalize()} terminates the Python process and frees up
the memory it was using.
}
\author{
Adapted from: Akitaka Matsuo
}
