% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_0_text_install.R
\name{textrpp_install}
\alias{textrpp_install}
\alias{textrpp_install_virtualenv}
\title{Install text required python packages in conda or virtualenv environment}
\usage{
textrpp_install(
  conda = "auto",
  rpp_version = c("torch==0.4.1", "transformers==3.3.1", "numpy", "nltk"),
  python_version = "3.7",
  envname = "textrpp_condaenv",
  pip = TRUE,
  python_path = NULL,
  prompt = TRUE
)

textrpp_install_virtualenv(
  rpp_version = c("torch==0.4.1", "transformers==3.3.1", "numpy", "nltk"),
  python_version = "3.6",
  pip_version = "pip",
  python_path = NULL,
  prompt = TRUE
)
}
\arguments{
\item{conda}{character; path to conda executable. Default "auto" which
automatically find the path}

\item{rpp_version}{character; rpp:s  to install.}

\item{python_version}{character; determine Python version for condaenv
installation.}

\item{envname}{character; name of the conda-environment to install text required python packages.
Default is "textrpp_condaenv".}

\item{pip}{\code{TRUE} to use pip for installing rpp If \code{FALSE}, conda
package manager with conda-forge channel will be used for installing rpp.}

\item{python_path}{character; path to Python in virtualenv installation}

\item{prompt}{logical; ask whether to proceed during the installation}

\item{pip_version}{character;}
}
\description{
Install text required python packages (rpp) in a self-contained environment.  For macOS and Linux-based systems, this will
  also install Python itself via a "miniconda" environment, for
  \code{textrpp_install}.  Alternatively, an existing conda installation may be
  used, by specifying its path.  The default setting of \code{"auto"} will
  locate and use an existing installation automatically, or download and
  install one if none exists.

  For Windows, automatic installation of miniconda installation is not currently
  available, so the user will need to \href{https://conda.io/projects/conda/en/latest/user-guide/install/index.html}{miniconda (or Anaconda) manually}.

If you wish to install Python in a "virtualenv", use the
  \code{textrpp_install_virtualenv} function.
}
\examples{
\dontrun{
# install text required python packages in a miniconda environment (macOS and Linux)
textrpp_install(prompt = FALSE)

# install text required python packages to an existing conda environment
textrpp_install(conda = "~/anaconda/bin/")
}
\dontrun{
# install text required python packages in a virtualenv environment
textrpp_install_virtualenv()
}
}
