% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_0_text_install.R
\name{textrpp_uninstall}
\alias{textrpp_uninstall}
\title{Uninstall textrpp conda environment}
\usage{
textrpp_uninstall(conda = "auto", prompt = TRUE, envname = "textrpp_condaenv")
}
\arguments{
\item{conda}{path to conda executable, default to "auto" which automatically
finds the path}

\item{prompt}{logical; ask whether to proceed during the installation}

\item{envname}{character; name of conda environment to remove}
}
\description{
Removes the conda environment created by textrpp_install()
}
