% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_3_textSentiment.R
\name{textSentiment}
\alias{textSentiment}
\title{Predict label and score of a text using huggingface model}
\usage{
textSentiment(
  x,
  model,
  device = "cpu",
  tokenizer_parallelism = FALSE,
  logging_level = "error"
)
}
\arguments{
\item{x}{A character variable or a tibble/dataframe with at least one character variable.}

\item{model}{Character string specifying pre-trained language model (default 'bert-base-uncased').
For full list of options see pretrained models at
\href{https://huggingface.co/transformers/pretrained_models.html}{HuggingFace}.
For example use "...".}

\item{device}{Name of device to use: 'cpu', 'gpu', or 'gpu:k' where k is a specific device number}

\item{tokenizer_parallelism}{If TRUE this will turn on tokenizer parallelism. Default FALSE.}

\item{logging_level}{Set the logging level. Default: "warning".
Options (ordered from less logging to more logging): critical, error, warning, info, debug}
}
\value{
A tibble with predicted labels and scores for each text variable.
The comment of the object show the model-name and computation time.
}
\description{
Predict label and score of a text using huggingface model
}
\examples{
\donttest{
# x <- Language_based_assessment_data_8[1:2, 1:2]
# word_embeddings_with_layers <- textEmbedLayersOutput(x, layers = 11:12)
}
}
\seealso{
see \code{\link{textEmbedLayerAggregation}} and \code{\link{textEmbed}}
}
