% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_2_LSA.R
\name{textSpace}
\alias{textSpace}
\title{textSpace: Create an LSA-based space with semantic representation for each word}
\usage{
textSpace(x, dims = dimcalc_share())
}
\arguments{
\item{x}{column with text.}

\item{dims}{number of dimensions to create, where the default dimcalc_share() is a
function from lsa that assisst in selecting number of dimensions.}
}
\value{
A list with two spaces ("tk" and "df") with words in the first column, followed
by N number of columns representing the words' semantic representation.
}
\description{
textSpace: Create an LSA-based space with semantic representation for each word
}
\examples{
example_rawdata <- sq_data_tutorial8_10$harmonywords
space <- textSpace(example_rawdata)
}
\seealso{
see \code{\link{textStaticSpace}}
}
