% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_2_LSA.R
\name{textStaticSpace}
\alias{textStaticSpace}
\title{textStaticSpace: Applies semantic representations to all character variables}
\usage{
textStaticSpace(df, space, tk_df = "tk", aggregate = "mean")
}
\arguments{
\item{df}{dataframe with that at least contain one character column.}

\item{space}{semantic space from textSpace.}

\item{tk_df}{option to use either the "tk" of "df" space from textSpace.}

\item{aggregate}{method to aggregate semantic representation when their are more than a single word.
(default is "mean"; see also "min" and "max")}
}
\value{
A list with tibbles for each character variable. Each tibble comprises a column with the text, followed by
collumns representing the semantic representations of the text. The tibbles are called the same as the original variable.
}
\description{
textStaticSpace: Applies semantic representations to all character variables
}
\seealso{
see \code{\link{textSpace}}
}
