% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_5_textDomainCompare.R
\name{textTokenizeAndCount}
\alias{textTokenizeAndCount}
\title{Tokenize and count}
\usage{
textTokenizeAndCount(data, n_remove_threshold = 3)
}
\arguments{
\item{data}{(string) Language to tokenise and count.}

\item{n_remove_threshold}{(numeric) Threshold deciding which words to remove}
}
\value{
A word-frequency data frame (can be saved to a model object or compared in textDomainCompare).
}
\description{
Tokenize and count
}
\examples{
\dontrun{
textTokenizeAndCount(Language_based_assessment_data_8["harmonytexts"])
}
}
\seealso{
see \code{\link{textDomainCompare}}
}
