% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_1_textTopic.R
\name{textTopic}
\alias{textTopic}
\title{BERTopic implementation with prediction extraction (EXPERIMENTAL - saving training-data)}
\usage{
textTopic(
  data,
  variable_name,
  embedding_model = "distilroberta",
  umap_model = "default",
  hdbscan_model = "default",
  vectorizer_model = "default",
  representation_model = "mmr",
  num_top_words = 10,
  n_gram_range = c(1, 3),
  stopwords = "english",
  min_df = 10,
  bm25_weighting = FALSE,
  reduce_frequent_words = TRUE,
  seed = 2024,
  save_dir = "./results"
)
}
\arguments{
\item{data}{(data.frame) Data including one text-variable to derive topcis from (with optional numerci or categorical variables)
that can be used later in the pipeline for significance test topics according to these variables.}

\item{variable_name}{(string)  The variable name in data of the text-variable to derive the topics from}

\item{embedding_model}{(string) embedding model to use for embedding data, choose from miniLM, mpnet, multi-mpnet, distilroberta}

\item{umap_model}{(string) dimension reduction algorithm, currently only "default" supported}

\item{hdbscan_model}{(string) clustering algorihtm, currently only "default" supported}

\item{vectorizer_model}{(string) vectorizer model, currently only "default" supported}

\item{representation_model}{(string) representation models used for topics, "keybert" or "mmr"}

\item{num_top_words}{(int) determine the number of top words for each topic}

\item{n_gram_range}{(list) ngram range used for vectorizer model}

\item{stopwords}{(string)}

\item{min_df}{(int) minimum document frequency of terms}

\item{bm25_weighting}{(bool) determine whether bm25_weighting is used for ClassTfidfTransformer}

\item{reduce_frequent_words}{(bool) determine whether frequent words are reduced by ClassTfidfTransformer}

\item{seed}{(int) set random seed for intialization of umap model}

\item{save_dir}{(string) set directory for saving results, defaults to "./results"}
}
\value{
A folder containing the model, data, another folder with terms and values for each topic, document-topic matrix
}
\description{
BERTopic implementation with prediction extraction (EXPERIMENTAL - saving training-data)
}
