% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_1_textTopics.R
\name{textTopicsReduce}
\alias{textTopicsReduce}
\title{textTopicsReduce (EXPERIMENTAL)}
\usage{
textTopicsReduce(
  data,
  data_var,
  n_topics = 10,
  load_path = "./results",
  save_dir = "./results_reduced",
  embedding_model = "default"
)
}
\arguments{
\item{data}{(tibble/data.frame) A tibble with a text-variable to be analysed, and optional
numeric/categorical variables that you might want to use for later analyses testing the
significance of topics in relation to these variables.}

\item{data_var}{(string)  Name of the text-variable in the data tibble that you want
to perform topic modeling on.}

\item{n_topics}{(string) The dimension reduction algorithm, currently only "default"
is supported.}

\item{load_path}{(string) The clustering algorithm to use, currently only "default"
is supported.}

\item{save_dir}{(string) The directory for saving results.}

\item{embedding_model}{(string) Name of the embedding model to use such as "miniLM", "mpnet",
"multi-mpnet", "distilroberta".}
}
\value{
A folder containing the model, data, folder with terms and values for each topic,
and the document-topic matrix. Moreover the model itself is returned formatted as a data.frame
together with metdata.
}
\description{
textTopicsReduce (EXPERIMENTAL)
}
\seealso{
See \code{\link{textTopics}} \code{\link{textTopicsTest}} and \code{\link{textTopicsWordcloud}}.
}
