% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_2_textTopicsAnalysis.R
\name{textTopicsTest}
\alias{textTopicsTest}
\title{Wrapper for topicsTest function from the topics package}
\usage{
textTopicsTest(
  model,
  pred_var_x = NULL,
  pred_var_y = NULL,
  group_var = NULL,
  control_vars = c(),
  test_method = "linear_regression",
  p_alpha = 0.05,
  p_adjust_method = "fdr",
  load_dir = NULL
)
}
\arguments{
\item{model}{(list) The trained model}

\item{pred_var_x}{(string) The x variable name to be predicted, and to be plotted (only needed for regression or correlation)}

\item{pred_var_y}{(string) The y variable name to be predicted, and to be plotted (only needed for regression or correlation)}

\item{group_var}{(string) The variable to group by (only needed for t-test)}

\item{control_vars}{(vector) The control variables (not supported yet)}

\item{test_method}{(string) The test method to use, either "correlation","t-test", "linear_regression","logistic_regression", or "ridge_regression"}

\item{p_alpha}{(numeric) Threshold of p value set by the user for visualising significant topics}

\item{p_adjust_method}{(character) Method to adjust/correct p-values for multiple comparisons
(default = "none"; see also "holm", "hochberg", "hommel", "bonferroni", "BH", "BY",  "fdr").}

\item{load_dir}{(string) The directory to load the test from, if NULL, the test will not be loaded}
}
\value{
A list of the test results, test method, and prediction variable
}
\description{
Wrapper for topicsTest function from the topics package
}
