% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_1_textTopics.R
\name{textTopicsTree}
\alias{textTopicsTree}
\title{textTopicsTest (EXPERIMENTAL) to get the hierarchical topic tree}
\usage{
textTopicsTree(topic_model, data, data_var)
}
\arguments{
\item{topic_model}{(list) The output from textTopics.}

\item{data}{(tibble/data.frame) A tibble with the data}

\item{data_var}{(string) The name of the text variable that the topic model was trained on}
}
\value{
prints a hierarchical topic tree on the console
}
\description{
textTopicsTest (EXPERIMENTAL) to get the hierarchical topic tree
}
