% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_2_textTopicsAnalysis.R
\name{textTopicsWordcloud}
\alias{textTopicsWordcloud}
\title{Plot word clouds}
\usage{
textTopicsWordcloud(model = NULL, ngrams = NULL, test = NULL, seed = 2024, ...)
}
\arguments{
\item{model}{(list) A trained topics model. For examples from topicsModel(). Should be NULL if plotting ngrams.}

\item{ngrams}{(list) The output from the the topicsGram() function . Should be NULL if plotting topics.}

\item{test}{(list) The test results; if plotting according to dimension(s) include the object from topicsTest() function.}

\item{seed}{(integer) The seed to set for reproducibility; need to be the same seed number as in in}

\item{...}{Parameters from the topicsPlot() function in the topics package.}
}
\value{
The function saves figures in the save_dir.
}
\description{
This function create word clouds and topic fugures
}
