% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_2_textTopicsAnalysis.R
\name{textTopicTest}
\alias{textTopicTest}
\title{The function for topic testing}
\usage{
textTopicTest(
  model,
  preds,
  data,
  group_var,
  control_vars,
  test_method,
  seed,
  load_dir = NULL,
  save_dir = NULL
)
}
\arguments{
\item{model}{(data.frame) The model returned from textTopics()$model.}

\item{preds}{(data.frame) The preds returned from textTopics()$preds}

\item{data}{(data.frame) The data on which the topic model was trained on, returned from textTopics()$train_data}

\item{group_var}{(string) Variable for t-test, linear, binary or ridge regression}

\item{control_vars}{(list) Control variable for linear or binary regression}

\item{test_method}{(string) Choose between "correlation", "t-test", "binary_regression", "linear_regression" or "ridge_regression"}

\item{seed}{(int) random seed}

\item{load_dir}{(string) if specified, the function returns the precomputed analysis from the directory, otherwise leave blank}

\item{save_dir}{(string) save analysis in specified directory, if left blank, analysis is not saved}
}
\value{
Test
}
\description{
The function for topic testing
}
