% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_1_textTrain.R
\name{textTrain}
\alias{textTrain}
\title{Train word embeddings to a numeric (ridge regression) or categorical (random forest) variable.}
\usage{
textTrain(x, y, force_train_method = "automatic", ...)
}
\arguments{
\item{x}{Word embeddings from textEmbed (or textLayerAggregation).}

\item{y}{Numeric variable to predict.}

\item{force_train_method}{default is "automatic", so if y is a factor random_forest is used, and if y is numeric ridge regression
is used. This can be overridden using "regression" or "random_forest".}

\item{...}{Arguments from textTrainRegression or textTrainRandomForest the textTrain function.}
}
\value{
A correlation between predicted and observed values; as well as a tibble of predicted values.
}
\description{
Train word embeddings to a numeric (ridge regression) or categorical (random forest) variable.
}
\examples{
wordembeddings <- wordembeddings4
ratings_data <- Language_based_assessment_data_8
results <- textTrain(wordembeddings$harmonytext,
                     ratings_data$hilstotal,
                     penalty = 1, # 1 due to computational constraints for the example context
                     multi_cores = FALSE
)
}
\seealso{
\code{\link{textTrainRegression}} \code{\link{textTrainRandomForest}}
\code{\link{textTrainLists}} \code{\link{textDiff}}
}
