% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_1_textTrain.R
\name{textTrainLists}
\alias{textTrainLists}
\title{textTrainLists trains word embeddings from several text variable to several numeric variable.}
\usage{
textTrainLists(
  x,
  y,
  trainMethod = "regression",
  nrFolds_k = 10,
  preProcessPCAthresh = 0.95,
  strata_y = "y",
  methodCor = "pearson",
  trees = 500,
  nrFolds_k_out = 10,
  ...
)
}
\arguments{
\item{x}{List of several Wordembeddings with same length from textImport (NB remove single_word_we).}

\item{y}{A Tibble with numeric variables to predict.}

\item{trainMethod}{Method to train wordembeddings, default "regression", or "randomForest"}

\item{nrFolds_k}{Number of folds to use.}

\item{preProcessPCAthresh}{Preprocessing threshold.}

\item{strata_y}{variables to stratify according; default y, can set to NULL}

\item{methodCor}{Type of correlation used in evaluation; default pearson (see also "spearman", "kendall").}

\item{trees}{Number of trees used in random forest.}

\item{nrFolds_k_out}{Number of outer folds (only for textTrainCVpredictions textTrainCVpredictionsRF).}

\item{...}{Arguments from textTrain.}
}
\value{
A correlation between predicted and observed values; as well as predicted values.
}
\description{
textTrainLists trains word embeddings from several text variable to several numeric variable.
}
\examples{
wordembeddings <- wordembeddings4_10[1:2]
ratings_data <- sq_data_tutorial4_10[1:2]
wordembeddings <- textTrainLists(wordembeddings, ratings_data, nrFolds_k = 2)
}
\seealso{
see \code{\link{textTrain}}
}
