% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_2_textTrainMulti.R
\name{textTrainMultiTexts}
\alias{textTrainMultiTexts}
\title{Train several word embeddings from several text variables to a numeric variable.}
\usage{
textTrainMultiTexts(
  xlist,
  y,
  preProcessThresh = 0.95,
  nrFolds_k = 10,
  strata_y = "y",
  methodCor = "pearson",
  describe_model = "Describe the model further and share it with others"
)
}
\arguments{
\item{xlist}{list of word embeddings from textEmbed.}

\item{y}{Numeric variable to predict.}

\item{preProcessThresh}{Preprocessing threshold for amount of variance to retain (default 0.95).}

\item{nrFolds_k}{Number of folds to use.}

\item{strata_y}{Variable to stratify according (default y; can set to NULL).}

\item{methodCor}{Type of correlation used in evaluation (default "pearson"; can set to "spearman" or "kendall").}

\item{describe_model}{Text to describe your model.}
}
\value{
A correlation between predicted and observed values; as well as predicted values.
}
\description{
Concenate word embeddings from several text variables to predict an outcome. The word embeddings for each text variable
go through seperate PCAs, where the PCA components are concenated and used in a ridge regression.
}
\examples{
wordembeddings <- wordembeddings4_10[1:4]
ratings_data <- Language_based_assessment_data_8_10$hilstotal
wordembeddings <- textTrainMultiTexts(wordembeddings, ratings_data, nrFolds_k = 2)
}
\seealso{
see \code{\link{textTrainLists}} \code{\link{textDiff}}
}
