% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_3_textTrainRandomForest.R
\name{textTrainRandomForest}
\alias{textTrainRandomForest}
\title{Train word embeddings to a categorical variable using random forrest.}
\usage{
textTrainRandomForest(
  x,
  y,
  outside_strata_y = "y",
  inside_strata_y = "y",
  mtry = c(1, 5, 10, 15, 30, 40),
  min_n = c(1, 5, 10, 15, 30, 40),
  trees = 1000,
  model_description = "Consider writing a description of your model here",
  multi_cores = TRUE
)
}
\arguments{
\item{x}{Word embeddings from textEmbed.}

\item{y}{Categorical variable to predict.}

\item{outside_strata_y}{Variable to stratify according (default "y"; can also set to NULL).}

\item{inside_strata_y}{Variable to stratify according (default "y"; can also set to NULL).}

\item{mtry}{hyper parameter that may be tuned;  default:c(1, 20, 40),}

\item{min_n}{hyper parameter that may be tuned; default: c(1, 20, 40)}

\item{trees}{Number of trees to use (default 1000).}

\item{model_description}{Text to describe your model (optional; good when sharing the model with others).}

\item{multi_cores}{If TRUE enables the use of multiple cores if computer/system allows for it (hence it can
make the analyses considerably faster to run).}
}
\value{
A chi-square result between predicted and observed values; as well as predicted and true values.
}
\description{
Train word embeddings to a categorical variable using random forrest.
}
\examples{
wordembeddings <- wordembeddings4_10
example_categories <- as.factor(c(1, 2, 1, 2, 1, 2, 1, 2, 1, 2))
results <- textTrainRandomForest(wordembeddings$harmonywords,
                                 example_categories,
                                 mtry  = c(1),       # this is short because of testing
                                 min_n = c(1),       # this is short because of testing
                                 multi_cores = FALSE # This is FALSE due to CRAN testing.
)
}
\seealso{
see \code{\link{textTrainLists}} \code{\link{textDiff}}
}
