% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_1_textTransform.R
\name{textTransform}
\alias{textTransform}
\title{textTransform extracts word embeddings for all text in character variables in dataframe}
\usage{
textTransform(
  data,
  pretrained_weights = "bert-base-uncased",
  tokenizer_class = BertTokenizer,
  model_class = BertModel,
  layers = "all",
  return_tokens = TRUE
)
}
\arguments{
\item{data}{Tibble.}

\item{pretrained_weights}{Character specifying pre-trained language model through RBERT. Current options
"bert_base_uncased", "bert_base_cased", "bert_large_uncased",
"bert_large_cased", "bert_large_uncased_wwm", "bert_large_cased_wwm",
"bert_base_multilingual_cased", "bert_base_chinese", "scibert_scivocab_uncased",
"scibert_scivocab_cased", "scibert_basevocab_uncased", and "scibert_basevocab_cased".}

\item{tokenizer_class}{tokenizer that match pretrained_weights}

\item{model_class}{model class that matches pretrained_weights and tokenizer_class}

\item{layers}{best to extract 'all', and then remove in other function, when aggregating embeddings}

\item{return_tokens}{f}
}
\value{
A tibble with tokens, layer identifyer and word embeddings. Note that layer 0 is the input embedding to the transformer
}
\description{
textTransform extracts word embeddings for all text in character variables in dataframe
}
\seealso{
see \code{\link{textTrain}}, \code{\link{textDiff}} and \code{\link{textSimilarity}}
}
