% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_textSimilarity_and_textTtest.R
\name{textTtest}
\alias{textTtest}
\title{textTtest computes the semantic similiarty between a character variable and a word norm (i.e., a text in one cell).}
\usage{
textTtest(x, y, nrFolds = 10, ...)
}
\arguments{
\item{x}{wordembeddings from textImport (with several rows of text).}

\item{y}{wordembeddings from textImport (with only one text).}

\item{nrFolds}{Number of folds used in the creation of semnatic Ttest scores.}

\item{...}{Arguments from t.test function.}
}
\value{
Results from a t-test including p-values, t-statistics and cohen's D.
}
\description{
textTtest computes the semantic similiarty between a character variable and a word norm (i.e., a text in one cell).
}
\examples{
wordembeddings <- wordembeddings4_10
Harmony_Satisfaction_ttest <- textTtest(wordembeddings$harmonytext, wordembeddings$satisfactiontext)
Harmony_Satisfaction_ttest
}
\seealso{
see \code{\link{textTrain}} and \code{\link{textSimilarity}}
}
