% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_1_textSimilarity_and_textTtest.R
\name{textTtestscores}
\alias{textTtestscores}
\title{textTtestscores computes the semantic similiarty between a character variable and a word norm (i.e., a text in one cell).}
\usage{
textTtestscores(x, y, nrFolds = 10)
}
\arguments{
\item{x}{Wordembeddings from textImport.}

\item{y}{Wordembeddings from textImport.}

\item{nrFolds}{Number of folds used.}
}
\value{
A vector with semantic similarity scores (based on cosine).
}
\description{
textTtestscores computes the semantic similiarty between a character variable and a word norm (i.e., a text in one cell).
}
\examples{
wordembeddings <- wordembeddings4_10
similiarty_scores <- textTtestscores(wordembeddings$harmonytext, wordembeddings$satisfactiontext)
}
\seealso{
see \code{\link{textTtest}}
}
