% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleDataDescription.R
\docType{data}
\name{word_embeddings_4}
\alias{word_embeddings_4}
\title{Word embeddings for 4 text variables for 40 participants}
\format{
A list with word embeddings for harmony words, satisfaction words, harmony text,
satisfaction text and decontextualized word embeddings. BERT-base embeddings based on
mean aggregation of layer 11 and 12.
\describe{
  \item{words}{words}
  \item{n}{word frequency}
  \item{Dim1:Dim768}{Word embeddings dimensions}
}
}
\source{
\url{https://osf.io/preprints/psyarxiv/er6t7/}
}
\usage{
word_embeddings_4
}
\description{
The dataset is a shortened version of the data sets of Study 3-5
from Kjell, Kjell, Garcia and Sikström 2018.
}
\keyword{datasets}
