% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleDataDescription.R
\docType{data}
\name{wordembeddings4_100}
\alias{wordembeddings4_100}
\title{Example text and numeric data.}
\format{A data frame with 608 rows and 4 variables:
\describe{
  \item{sequence_index}{output from textImport}
  \item{segment_index}{output from textImport}
  \item{token_index}{output from textImport}
  \item{layer_index}{output from textImport}
  \item{V1:V768}{dimension output from textImport}
  ...
}}
\source{
\url{https://psyarxiv.com/er6t7/}
}
\usage{
wordembeddings4_100
}
\description{
The dataset is a shortened version of the data sets of Study 3-5
from Kjell, Kjell, Garcia and Sikström 2018.
}
\keyword{datasets}
