% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleDataDescription.R
\docType{data}
\name{centrality_data_harmony}
\alias{centrality_data_harmony}
\title{Example data for plotting a Semantic Centrality Plot.}
\format{
A data frame with 2,146 and 4 variables:
\describe{
  \item{words}{output from textProjectionData}
  \item{n}{output from textProjectionData}
  \item{central_cosine}{output from textProjectionData}
  \item{n_percent}{output from textProjectionData}
  ...
}
}
\source{
\url{https://psyarxiv.com/er6t7/}
}
\usage{
centrality_data_harmony
}
\description{
The dataset is a shortened version of the data sets of Study 1
from Kjell, et al., 2016.
}
\keyword{datasets}
