% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleDataDescription.R
\docType{data}
\name{sq_data_plottingHw_HILSSSWLS_100}
\alias{sq_data_plottingHw_HILSSSWLS_100}
\title{Example data for plotting.}
\format{A data frame with 43 rows and 13 variables:
\describe{
  \item{words}{output from textPlotData}
  \item{mean.x}{output from textPlotData}
  \item{n.x}{output from textPlotData}
  \item{sd.x}{output from textPlotData}
  \item{p_values.x}{output from textPlotData}
  \item{t_statistic.x}{output from textPlotData}
  \item{cohensD.x}{output from textPlotData}
   \item{mean.y}{output from textPlotData}
  \item{n.y}{output from textPlotData}
  \item{sd.y}{output from textPlotData}
  \item{p_values.y}{output from textPlotData}
  \item{t_statistic.y}{output from textPlotData}
  \item{cohensD.y}{output from textPlotData}
  ...
}}
\source{
\url{https://psyarxiv.com/er6t7/}
}
\usage{
sq_data_plottingHw_HILSSSWLS_100
}
\description{
The dataset is a shortened version of the data sets of Study 3-5
from Kjell, Kjell, Garcia and Sikström 2018.
}
\keyword{datasets}
