% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_1_textPlot.R
\name{textCentralityPlot}
\alias{textCentralityPlot}
\title{textCentralityPlot plots words}
\usage{
textCentralityPlot(
  word_data,
  min_freq_words = 1,
  plot_n_word_extreme = 10,
  plot_n_word_frequency = 10,
  plot_n_words_middle = 10,
  titles_color = "#61605e",
  x_axes = "central_cosine",
  title_top = "Semantic Centrality Plot",
  x_axes_label = "Semantic Centrality",
  y_axes_values = element_blank(),
  scale_x_axes_lim = NULL,
  scale_y_axes_lim = NULL,
  word_font = NULL,
  centrality_color_codes = c("#EAEAEA", "#85DB8E", "#398CF9"),
  word_size_range = c(3, 8),
  position_jitter_hight = 0,
  position_jitter_width = 0.03,
  point_size = 0.5,
  arrow_transparency = 0.1,
  points_without_words_size = 0.5,
  points_without_words_alpha = 0.5,
  legend_title = "SC",
  legend_x_axes_label = "x",
  legend_x_position = 0.02,
  legend_y_position = 0.02,
  legend_h_size = 0.2,
  legend_w_size = 0.2,
  legend_title_size = 7,
  legend_number_size = 2
)
}
\arguments{
\item{word_data}{Dataframe from textPlotData.}

\item{min_freq_words}{Select words to signficance test that has occured at least min_freq_words (default = 1).}

\item{plot_n_word_extreme}{Number of words that are extreme on dot product projection per dimension.
(i.e., even if not significant;  per dimensions, where duplicates are removed).}

\item{plot_n_word_frequency}{Number of words based on being most frequent.
(i.e., even if not significant).}

\item{plot_n_words_middle}{Number of words plotted that are in the middle in dot product projection score
(i.e., even if not significant;  per dimensions, where duplicates are removed).}

\item{titles_color}{Color for all the titles (default: "#61605e")}

\item{x_axes}{Variable to be plotted on the x axes (default is "dot_product.x").}

\item{title_top}{string for title (default "  ")}

\item{x_axes_label}{Label on the x-axes.}

\item{y_axes_values}{NULL,}

\item{scale_x_axes_lim}{Manually set the legth of the x-axes (deafult=NULL; e.g., try c(-5, 5)).}

\item{scale_y_axes_lim}{Manually set the legth of the y-axes (deafult=NULL; e.g., try c(-5, 5)).}

\item{word_font}{Type of font (default: "Arial").}

\item{centrality_color_codes}{The diffent colors of the words (default: c("#EAEAEA","#85DB8E", "#398CF9")).}

\item{word_size_range}{a vector with minimum and maximum (default: c(3, 8)).}

\item{position_jitter_hight}{degree of jitter hight (default: .0).}

\item{position_jitter_width}{degree of jitter hight (default: .03).}

\item{point_size}{the size of the points indicating the words' position (default: 0.5).}

\item{arrow_transparency}{the transparency of the lines between each word and point (default: 0.1).}

\item{points_without_words_size}{sets the size of the points (without showing words, default is to not show it, i.e., 0).}

\item{points_without_words_alpha}{sets the transperency of the points (without showing words, default is to not show it, i.e., 0).}

\item{legend_title}{Label on the color legend (default: "(DPP)".}

\item{legend_x_axes_label}{Label on the color legend (default: "(x)".}

\item{legend_x_position}{Position on the x coordinates of the color legend (default: 0.02).}

\item{legend_y_position}{Position on the y coordinates of the color legend (default: 0.05).}

\item{legend_h_size}{The height of the color legend (default 0.15).}

\item{legend_w_size}{The width of the color legend (default 0.15).}

\item{legend_title_size}{=7}

\item{legend_number_size}{size of the values in the legend (default = 2)}
}
\value{
A 1-dimensioal word plot.
}
\description{
textCentralityPlot plots words
}
\examples{
# The test-data included in the package is called: centrality_data_harmony
names(centrality_data_harmony)
# Plot
centrality_plot <- textCentralityPlot(
 word_data=centrality_data_harmony,
 min_freq_words = 10,
 plot_n_word_extreme = 10,
 plot_n_word_frequency = 10,
 plot_n_words_middle = 10,
 titles_color = "#61605e",
 x_axes = "central_cosine",

 title_top = "Semantic Centrality Plot",
 x_axes_label = "Semantic Centrality",

 word_font = NULL,
 centrality_color_codes = c("#EAEAEA","#85DB8E", "#398CF9"),
 word_size_range = c(3, 8),
 point_size = 0.5,
 arrow_transparency = 0.1,
 points_without_words_size = 0.5,
 points_without_words_alpha = 0.5,
)
centrality_plot


}
\seealso{
see \code{\link{textCentralityData}} and \code{\link{textProjectionData}}
}
