% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_1_textEmbed.R
\name{textHuggingFace}
\alias{textHuggingFace}
\title{Extract layers of hidden states (word embeddings) for all character variables in the dataframe}
\usage{
textHuggingFace(
  x,
  contexts = TRUE,
  decontexts = TRUE,
  pretrained_weights = "bert-base-uncased",
  tokenizer_class = BertTokenizer,
  model_class = BertModel,
  layers = "all",
  return_tokens = TRUE
)
}
\arguments{
\item{x}{Tibble/dataframe with at least one character variable.}

\item{contexts}{Provide word embeddings based on word contexts (standard method; default = TRUE).}

\item{decontexts}{Provide word embeddings of single words as input (embeddings used for plotting; default = TRUE).}

\item{pretrained_weights}{Character specifying pre-trained language model (default 'bert-base-uncased'). For an up-to-date
list of avaiable models see https://github.com/huggingface/transformers.}

\item{tokenizer_class}{Tokenizer that match pretrained_weights (default BertTokenizer). For an up-to-date
list of avaiable models see https://github.com/huggingface/transformers.}

\item{model_class}{Model class that matches pretrained_weights and tokenizer_class (default BertModel). For an up-to-date
list of avaiable models see https://github.com/huggingface/transformers.}

\item{layers}{Specify the layers that should be extracted (default 'all'). It is more efficient to only extract the layers
that you need (e.g., 11:12). Layer 0 is the decontextualised input layer (i.e., not comprising hidden states) and thus adviced to not use.
These layers can then be aggregated in the textLayerAggregation function.}

\item{return_tokens}{If TRUE, provide the tokens used in the specified transformer model.}
}
\value{
A tibble with tokens, layer identifyer and word embeddings. Note that layer 0 is the input embedding to the transformer, and should
normally not be used.
}
\description{
Extract layers of hidden states (word embeddings) for all character variables in the dataframe
}
\examples{
\dontrun{
x <- Language_based_assessment_data_8_10[1:2, 1:2]
wordembeddings <- textHuggingFace(x, layers = 'all')
}
}
\seealso{
see \code{\link{textLayerAggregation}} and \code{\link{textEmbed}}
}
