% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_textImport.R
\name{textImport}
\alias{textImport}
\title{textImport extracts word embeddings for all text in character variables in dataframe}
\usage{
textImport(
  x,
  layer_indexes_REBERT = 12,
  batch_size_IBT = 2L,
  token_index_IBT = 1,
  layer_index_IBT = 12,
  ...
)
}
\arguments{
\item{x}{Tibble.}

\item{layer_indexes_REBERT}{layers to be used from RBERT (default 12; in REBERT it is 1:12).}

\item{batch_size_IBT}{batch size from RBERT (default 2L)}

\item{token_index_IBT}{from RBERT}

\item{layer_index_IBT}{from RBERT}

\item{...}{arguments from RBERT function extract_features}
}
\value{
A list with word embeddings.
}
\description{
textImport extracts word embeddings for all text in character variables in dataframe
}
\examples{
x <- sq_data_tutorial8_10
library(dplyr)
wordembeddings <- textImport(x)
}
\seealso{
see \code{\link{textTrain}}, \code{\link{textTtest}} and \code{\link{textSimilarity}}
}
