% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_1_textEmbed.R
\name{textLayerAggregation}
\alias{textLayerAggregation}
\title{Select and aggregate layers of hidden states to form a word embeddings.}
\usage{
textLayerAggregation(
  word_embeddings_layers,
  layers = 1:12,
  aggregation = "mean",
  tokens_select = NULL,
  tokens_deselect = NULL
)
}
\arguments{
\item{word_embeddings_layers}{Layers outputted from textHuggingFace.}

\item{layers}{The numbers of the layers to be aggregated (e.g., c(11:12) to aggregate the eleventh and twelth; 'all' is not possible here).
Note that layer 0 is the input embedding to the transformer, and should normally not be used.}

\item{aggregation}{Method to aggregate the dimensions of each selected layer (default "mean"; see also "min" or "max).}

\item{tokens_select}{Option to only select embeddings linked to specifc token such as [CLS] and [SEP] (default NULL).}

\item{tokens_deselect}{Option to deselect embeddings linked to specifc token such as [CLS] and [SEP] (default NULL).}
}
\value{
A tibble with word embeddings. Note that layer 0 is the input embedding to the transformer, which is normally not used.
}
\description{
Select and aggregate layers of hidden states to form a word embeddings.
}
\examples{
\dontrun{
wordembeddings <- textLayerAggregation(word_embeddings_layers)
}
}
\seealso{
see \code{\link{textHuggingFace}} and \code{\link{textEmbed}}
}
