% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_1_textEmbed.R
\name{textModels}
\alias{textModels}
\title{Check downloaded, available models.}
\usage{
textModels()
}
\value{
List of names of models and tokenizers
}
\description{
Check downloaded, available models.
}
\examples{
\donttest{
textModel()
}
}
\seealso{
see \code{\link{textModelsRemove}}
}
