% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_3_textModels.R
\name{textModelsRemove}
\alias{textModelsRemove}
\title{Delete a specified model}
\usage{
textModelsRemove(target_model)
}
\arguments{
\item{target_model}{(string) The name of the model to be deleted.}
}
\value{
Confirmation whether the model has been deleted.
}
\description{
This functions delete specified mode and associated files.
}
\examples{
\dontrun{
textModelsRemove("name-of-model-to-delete")
}
}
\seealso{
see \code{\link{textModels}}
}
