% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_3_textModels.R
\name{textModelsRemove}
\alias{textModelsRemove}
\title{Delete a specified model and model associated files.}
\usage{
textModelsRemove(target_model)
}
\arguments{
\item{target_model}{Character string of the model name that you want to delete.}
}
\value{
Confirmation whether the model has been deleted.
}
\description{
Delete a specified model and model associated files.
}
\examples{
\dontrun{
textModelsRemove("name-of-model-to-delete")
}
}
\seealso{
see \code{\link{textModels}}
}
