% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_3_textNER.R
\name{textNER}
\alias{textNER}
\title{Named Entity Recognition STILL UNDER DEVELOPMENT
Uses any Model for token classification.
for predicting the classes of tokens in a sequence: person, organisation, location or miscellaneous).}
\usage{
textNER(
  x,
  model = "dslim/bert-base-NER",
  device = "cpu",
  tokenizer_parallelism = FALSE,
  logging_level = "warning",
  return_incorrect_results = FALSE
)
}
\arguments{
\item{x}{(string)  A character variable or a tibble/dataframe with at least one character variable.}

\item{model}{(string)  Specify a pre-trained language model that have been fine-tuned on a token classification task
(e.g., see "dslim/bert-base-NER").}

\item{device}{(string)  Name of device to use: 'cpu', 'gpu', or 'gpu:k' where k is a specific device number}

\item{tokenizer_parallelism}{(boolean)  If TRUE this will turn on tokenizer parallelism. Default FALSE.}

\item{logging_level}{(string)  Set the logging level. Default: "warning".
Options (ordered from less logging to more logging): critical, error, warning, info, debug}

\item{return_incorrect_results}{(boolean)  Many models are not created to be able to provide classifications - this setting
stops them from returning incorrect results.}
}
\value{
A tibble with.
}
\description{
Named Entity Recognition STILL UNDER DEVELOPMENT
Uses any Model for token classification.
for predicting the classes of tokens in a sequence: person, organisation, location or miscellaneous).
}
\examples{
\donttest{
ner_example <- textNER("The meaning of life is")
ner_example
}
}
\seealso{
see \code{\link{textEmbedLayerAggregation}} and \code{\link{textEmbed}}
}
