% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_textPlot.R
\name{textPlotData}
\alias{textPlotData}
\title{textPlotData computes variables about individual words for plotting.}
\usage{
textPlotData(
  words,
  wordembeddings,
  single_wordembeddings = single_wordembeddings_df,
  x,
  y = NULL,
  Bonferroni = TRUE,
  nrFolds = 10
)
}
\arguments{
\item{words}{Words to be plotted.}

\item{wordembeddings}{Wordembeddings from textImport for the words to be plotted .}

\item{single_wordembeddings}{Wordembeddings from textImport for individual words "}

\item{x}{Numeric variable of a dimension the words should be plotted according to (y=NULL)}

\item{y}{Numeric variable of a dimension the words should be plotted according to (y=NULL)}

\item{Bonferroni}{Whether to apply Bonferroni correction in relation to the number of t-test carried out for each individual word (default = TRUE)}

\item{nrFolds}{Number of folds used in creation of semantic similiarty scores for individual words.}
}
\value{
A dataframe with variables (e.g., including p-values, frequency, Cohen's D) for the individual words
that is used for the plotting in the textPlotViz function.
}
\description{
textPlotData computes variables about individual words for plotting.
}
\examples{
wordembeddings <- wordembeddings4_10
data <- sq_data_tutorial8_10
df_for_plotting <- textPlotData(data$harmonywords, wordembeddings$harmonywords,
wordembeddings$singlewords_we,
data$hilstotal, data$swlstotal,
Bonferroni = TRUE, nrFolds=2)
df_for_plotting
}
\seealso{
see \code{\link{textPlotViz}}
}
