% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_textPlot.R
\name{textPlotViz}
\alias{textPlotViz}
\title{textPlotViz trains word embeddings to a numeric variable.}
\usage{
textPlotViz(
  word_data,
  plot_n_words = 25,
  title_top = " ",
  titles_color = "#61605e",
  x_axes = "cohensD.x",
  y_axes = "cohensD.y",
  x_axes_label = "Cohen's D",
  y_axes_label = "Cohen's D",
  scale_x_axes_lim = NULL,
  scale_y_axes_lim = NULL,
  y_axes_values = NULL,
  word_font = "Helvetica",
  colors_words = c("#ff0000", "#ff8080", "white", "#99e699", "#33cc33"),
  colors_words_scale = c(-0.1, -0.01, 0, 0.01, 0.1),
  word_size_range = c(3, 8),
  position_jitter_hight = 0,
  position_jitter_width = 0.03,
  point_size = 0.5,
  arrow_transparency = 0.1
)
}
\arguments{
\item{word_data}{Dataframe from textPlotData.}

\item{plot_n_words}{Number of significant words to plot; selects fist according to lowest p-value and then to highest absolut Cohen's D (default 25 for each scale).}

\item{title_top}{string for title (default "  ")}

\item{titles_color}{Color for all the titles (default: "#61605e")}

\item{x_axes}{Variable to be plotted on the x axes (default is "cohensD.x").}

\item{y_axes}{Variable to be plotted on the y axes (default is "cohensD.y"). To only print 1-dimension insert NULL.}

\item{x_axes_label}{Label on the x-axes (default: "Cohen's D").}

\item{y_axes_label}{Label on the y-axes (default: "Cohen's D").}

\item{scale_x_axes_lim}{NULL,}

\item{scale_y_axes_lim}{NULL,}

\item{y_axes_values}{NULL,}

\item{word_font}{Type of font (default: "Helvetica").}

\item{colors_words}{The diffent colors of the words (default: c("#ff0000", "#ff8080", "white", "#1aff1a", "#00cc00")).}

\item{colors_words_scale}{Limits indicating where the colors should change (default: c(-0.1, -0.01, 0, 0.01, 0.1).}

\item{word_size_range}{a vector with minimum and maximum (default: c(3, 8)).}

\item{position_jitter_hight}{degree of jitter hight (default: .0).}

\item{position_jitter_width}{degree of jitter hight (default: .03).}

\item{point_size}{the size of the points indicating the words' position (default: 0.5).}

\item{arrow_transparency}{The transparency of the lines between each word and point (default: 0.1)}
}
\value{
A 1- or 2-dimensional word plot.
}
\description{
textPlotViz trains word embeddings to a numeric variable.
}
\examples{
# The test-data included in the package is called: sq_data_plottingHw_HILSSSWLS_100
plot <- textPlotViz(word_data = sq_data_plottingHw_HILSSSWLS_100,
     plot_n_words = 25,
     x_axes = "cohensD.x",
     y_axes = "cohensD.y",
     x_axes_label = "HILS: Cohen's D",
     y_axes_label = "SWLS: Cohen's D",
     scale_x_axes_lim = NULL,
     scale_y_axes_lim = NULL,
     y_axes_values = NULL,
     word_font = "Helvetica",
     colors_words = c("#ff0000", "#ff8080", "white", "#1aff1a", "#00cc00"),
     colors_words_scale = c(-0.1, -0.01, 0, 0.01, 0.1),
     word_size_range = c(3, 8),
     position_jitter_hight = .0,
     position_jitter_width = .03,
     point_size = 0.5,
     arrow_transparency = 0.1)
plot
}
\seealso{
see \code{\link{textPlotData}}
}
