% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_4_textPredict.R
\name{textPredict}
\alias{textPredict}
\title{Predict scores or classification from, e.g., textTrain.}
\usage{
textPredict(
  model_info,
  word_embeddings,
  x_append = NULL,
  type = NULL,
  dim_names = TRUE,
  ...
)
}
\arguments{
\item{model_info}{Model info (e.g., saved output from textTrain, textTrainRegression or textRandomForest).}

\item{word_embeddings}{Word embeddings}

\item{x_append}{Variables to be appended after the word embeddings (x).}

\item{type}{Type of prediction; e.g., "prob", "class".}

\item{dim_names}{Account for specific dimension names from textEmbed
(rather than generic names including Dim1, Dim2 etc.).}

\item{...}{Setting from stats::predict can be called.}
}
\value{
Predicted scores from word embeddings.
}
\description{
Predict scores or classification from, e.g., textTrain.
}
\examples{
word_embeddings <- word_embeddings_4
ratings_data <- Language_based_assessment_data_8
}
\seealso{
see \code{\link{textTrain}} \code{\link{textTrainLists}}
\code{\link{textTrainRandomForest}} \code{\link{textSimilarityTest}}
}
