% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_1_textTrain.R
\name{textPredict}
\alias{textPredict}
\title{Predict scores or classification from, e.g., textTrain.}
\usage{
textPredict(model_info = NULL, new_data = NULL)
}
\arguments{
\item{model_info}{model info (e.g., saved output from textTrain, textTrainRegression or textRandomForest).}

\item{new_data}{Word embeddings from new data to be predicted from.}
}
\value{
Predicted scores from word embeddings.
}
\description{
Predict scores or classification from, e.g., textTrain.
}
\examples{
wordembeddings <- wordembeddings4
ratings_data <- Language_based_assessment_data_8

}
\seealso{
see \code{\link{textEmbedLayerAggreation}} \code{\link{textTrainLists}}
\code{\link{textTrainRandomForest}} \code{\link{textSimilarityTest}}
}
