% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_1_textPlot.R
\name{textProjectionPlot}
\alias{textProjectionPlot}
\title{Plot words according to Dot Product Projections.}
\usage{
textProjectionPlot(
  word_data,
  k_n_words_two_test = FALSE,
  min_freq_words = 1,
  plot_n_words_square = 3,
  plot_n_words_p = 5,
  plot_n_word_extreme = 5,
  plot_n_word_frequency = 5,
  plot_n_words_middle = 5,
  titles_color = "#61605e",
  x_axes = "dot.x",
  y_axes = NULL,
  p_values_x = "p_values_dot.x",
  p_values_y = NULL,
  p_alpha = 0.05,
  p_adjust_method = "none",
  title_top = "Dot Product Projection",
  x_axes_label = "Dot product projection (DPP)",
  y_axes_label = "Dot product projection (DPP)",
  scale_x_axes_lim = NULL,
  scale_y_axes_lim = NULL,
  y_axes_values_hide = TRUE,
  word_font = NULL,
  bivariate_color_codes = c("#398CF9", "#60A1F7", "#5dc688", "#e07f6a", "#EAEAEA",
    "#40DD52", "#FF0000", "#EA7467", "#85DB8E"),
  word_size_range = c(3, 8),
  position_jitter_hight = 0,
  position_jitter_width = 0.03,
  point_size = 0.5,
  arrow_transparency = 0.1,
  points_without_words_size = 0.2,
  points_without_words_alpha = 0.2,
  legend_title = "DPP",
  legend_x_axes_label = "x",
  legend_y_axes_label = "y",
  legend_x_position = 0.02,
  legend_y_position = 0.02,
  legend_h_size = 0.2,
  legend_w_size = 0.2,
  legend_title_size = 7,
  legend_number_size = 2
)
}
\arguments{
\item{word_data}{Dataframe from textProjectionData}

\item{k_n_words_two_test}{Select the k most frequent words to significance test (k = sqrt(100*N); N = number of paricipant responses). Default = TRUE.}

\item{min_freq_words}{Select words to signficance test that have occurred at least min_freq_words (default = 1).}

\item{plot_n_words_square}{Select number of significant words in each square of the figure to plot.}

\item{plot_n_words_p}{Number of significant words to plot (n per x-axes and n per y-axes, where duplicates are removed);
selects fist according to lowest p-value and then to frequency.}

\item{plot_n_word_extreme}{Number of words that are extreme on dot product projection per dimension.
(i.e., even if not significant; per dimensions, where duplicates are removed).}

\item{plot_n_word_frequency}{Number of words based on being most frequent.
(i.e., even if not significant).}

\item{plot_n_words_middle}{Number of words plotted that are in the middle in dot product projection score
(i.e., even if not significant;  per dimensions, where duplicates are removed).}

\item{titles_color}{Color for all the titles (default: "#61605e")}

\item{x_axes}{Variable to be plotted on the x-axes (default is "dot_product.x").}

\item{y_axes}{Variable to be plotted on the y-axes (default is "cohensD.y"). To only print 1-dimension insert NULL.}

\item{p_values_x}{P-value variable to plot according to.}

\item{p_values_y}{P-value variable to plot according to.}

\item{p_alpha}{Alpha (default = .05).}

\item{p_adjust_method}{Method to adjust/correct p-values for multiple comparisons (deafult = "holm"; see also "none", "hochberg",
"hommel", "bonferroni", "BH", "BY",  "fdr").}

\item{title_top}{Title (default "  ")}

\item{x_axes_label}{Label on the x-axes.}

\item{y_axes_label}{Label on the y-axes.}

\item{scale_x_axes_lim}{Manually set the length of the x-axes (deafult = NULL, which uses ggplot2::scale_x_continuous(limits = scale_x_axes_lim); change e.g., by trying c(-5, 5)).}

\item{scale_y_axes_lim}{Manually set the length of the y-axes (deafult = NULL; which uses uses ggplot2::scale_y_continuous(limits = scale_y_axes_lim); change e.g., by trying c(-5, 5)).}

\item{y_axes_values_hide}{If TRUE, hides the values on y-axes (which is good for one-dimensional plots).}

\item{word_font}{Font type (default: "Arial").}

\item{bivariate_color_codes}{The diffent colors of the words (default: c("#398CF9", "#60A1F7", "#5dc688",
"#e07f6a", "#EAEAEA", "#40DD52",
"#FF0000", "#EA7467", "#85DB8E")).}

\item{word_size_range}{Vector with minimum and maximum font size (default: c(3, 8)).}

\item{position_jitter_hight}{Jitter hight (default: .0).}

\item{position_jitter_width}{Jitter width (default: .03).}

\item{point_size}{Size of the points indicating the words' position (default: 0.5).}

\item{arrow_transparency}{Transparency of the lines between each word and point (default: 0.1).}

\item{points_without_words_size}{Size of the points not linked with a words (default is to not show it, i.e., 0).}

\item{points_without_words_alpha}{Transperency of the points not linked with a words (default is to not show it, i.e., 0).}

\item{legend_title}{Title on the color legend (default: "(DPP)".}

\item{legend_x_axes_label}{Label on the color legend (default: "(x)".}

\item{legend_y_axes_label}{Label on the color legend (default: "(y)".}

\item{legend_x_position}{Position on the x coordinates of the color legend (default: 0.02).}

\item{legend_y_position}{Position on the y coordinates of the color legend (default: 0.05).}

\item{legend_h_size}{Height of the color legend (default 0.15).}

\item{legend_w_size}{Width of the color legend (default 0.15).}

\item{legend_title_size}{Font size (default: 7).}

\item{legend_number_size}{Font size of the values in the legend (default: 2).}
}
\value{
A 1- or 2-dimensional word plot.
}
\description{
Plot words according to Dot Product Projections.
}
\examples{
# The test-data included in the package is called: DP_projections_HILS_SWLS_100

# Dot Product Projection Plot
plot_projection <- textProjectionPlot(
 word_data = DP_projections_HILS_SWLS_100,
 k_n_words_two_test = FALSE,
 min_freq_words = 1,
 plot_n_words_square = 3,
 plot_n_words_p = 3,
 plot_n_word_extreme = 1,
 plot_n_word_frequency = 1,
 plot_n_words_middle = 1,
 x_axes = "dot.x",
 y_axes = "dot.y",
 p_values_x = "p_values_dot.x",
 p_values_y = "p_values_dot.y",
 p_alpha = 0.05,
 title_top = " Dot Product Projection (DPP)",
 x_axes_label = "Low vs. High HILS score",
 y_axes_label = "Low vs. High SWLS score",
 p_adjust_method = "bonferroni",
 scale_y_axes_lim = NULL
)
plot_projection

names(DP_projections_HILS_SWLS_100)

}
\seealso{
see \code{\link{textProjectionData}}
}
