% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_1_textSimilarity.R
\name{textSimilarity}
\alias{textSimilarity}
\title{Compute the cosine semantic similarity between two text variables.}
\usage{
textSimilarity(x, y)
}
\arguments{
\item{x}{Word embeddings from textEmbed.}

\item{y}{Word embeddings from textEmbed.}
}
\value{
A vector comprising cosine semantic similarity scores.
}
\description{
Compute the cosine semantic similarity between two text variables.
}
\examples{
library(dplyr)
wordembeddings <- wordembeddings4
similiarty_scores <- textSimilarity(wordembeddings$harmonytext, wordembeddings$satisfactiontext)
}
\seealso{
see \code{\link{textSimilarityNorm}} and \code{\link{textDiff}}
}
