% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_1_textSimilarity.R
\name{textSimilarity}
\alias{textSimilarity}
\title{textSimilarity computes the semantic similiarty between texts.}
\usage{
textSimilarity(x, y)
}
\arguments{
\item{x}{Wordembeddings from textEmbed.}

\item{y}{Wordembeddings from textEmbed.}
}
\value{
A vector with semantic similarity scores (based on cosine).
}
\description{
textSimilarity computes the semantic similiarty between texts.
}
\examples{
library(dplyr)
wordembeddings <- wordembeddings4_10
similiarty_scores <- textSimilarity(wordembeddings$harmonytext, wordembeddings$satisfactiontext)
}
\seealso{
see \code{\link{textSimilarityNorm}} and \code{\link{textDiff}}
}
