% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_1_textSimilarity.R
\name{textSimilarityMatrix}
\alias{textSimilarityMatrix}
\title{Compute semantic similarity scores between all combinations in a word embedding}
\usage{
textSimilarityMatrix(embedding)
}
\arguments{
\item{embedding}{Word embeddings (e.g., from textEmbed).}
}
\value{
A matrix of semantic similarity scores
}
\description{
Compute semantic similarity scores between all combinations in a word embedding
}
\examples{
similarity_scores <- textSimilarityMatrix(word_embeddings_4$harmonytext)
round(similarity_scores, 3)
}
\seealso{
see \code{\link{textSimilarityNorm}} and \code{\link{textSimilarityTest}}
}
