% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_1_textSimilarity.R
\name{textSimilarityNorm}
\alias{textSimilarityNorm}
\title{textSimilarityNorm computes the semantic similiarty between a character variable and a word norm (i.e., a text in one cell).}
\usage{
textSimilarityNorm(x, y)
}
\arguments{
\item{x}{Wordembeddings from textEmbed (with several rows of text).}

\item{y}{Wordembeddings from textEmbed (with only one text).}
}
\value{
A vector with semantic similarity scores (based on cosine).
}
\description{
textSimilarityNorm computes the semantic similiarty between a character variable and a word norm (i.e., a text in one cell).
}
\examples{
library(dplyr)
library(tibble)
harmonynorm <- c("harmony peace ")
satisfactionnorm <- c("satisfaction achievement")

norms <- tibble::tibble(harmonynorm, satisfactionnorm)
wordembeddings <- wordembeddings4_10
wordembeddings_wordnorm <- textEmbed(norms)
}
\seealso{
see \code{\link{textSimilarity}} and \code{\link{textDiff}}
}
