% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_1_textSimilarity_and_textTtest.R
\name{textSimilarityNorm}
\alias{textSimilarityNorm}
\title{textSimilarityNorm computes the semantic similiarty between a character variable and a word norm (i.e., a text in one cell).}
\usage{
textSimilarityNorm(x, y)
}
\arguments{
\item{x}{Wordembeddings from textImport (with several rows of text).}

\item{y}{Wordembeddings from textImport (with only one text).}
}
\value{
A vector with semantic similarity scores (based on cosine).
}
\description{
textSimilarityNorm computes the semantic similiarty between a character variable and a word norm (i.e., a text in one cell).
}
\examples{

harmonynorm <- c("harmony peace cooperation balance")
satisfactionnorm <- c("satisfaction achievement job fulfilled")
library(tibble)
norms <- tibble::tibble(harmonynorm, satisfactionnorm)
library(dplyr)
wordembeddings <- wordembeddings4_10
wordembeddings_wordnorm <- textImport(norms)
similiarty_scores <- textSimilarityNorm(
  wordembeddings$harmonytext,
  wordembeddings_wordnorm$harmonynorm
)
}
\seealso{
see \code{\link{textSimilarity}} and \code{\link{textTtest}}
}
