% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_1_textSimilarity.R
\name{textSimilarityNorm}
\alias{textSimilarityNorm}
\title{Compute the semantic similarity between a text variable and a word norm
(i.e., a text represented by one word embedding that represent a construct).}
\usage{
textSimilarityNorm(x, y, method = "cosine")
}
\arguments{
\item{x}{Word embeddings from textEmbed (with several rows of text).}

\item{y}{Word embedding from textEmbed (from only one text).}

\item{method}{Character string describing type of measure to be computed; default is "cosine" (see also
measures from textDistance (here computed as 1 - textDistance()) including "euclidean", "maximum", "manhattan", "canberra", "binary" and "minkowski").}
}
\value{
A vector comprising semantic similarity scores.
}
\description{
Compute the semantic similarity between a text variable and a word norm
(i.e., a text represented by one word embedding that represent a construct).
}
\examples{
\dontrun{
library(dplyr)
library(tibble)
harmonynorm <- c("harmony peace ")
satisfactionnorm <- c("satisfaction achievement")

norms <- tibble::tibble(harmonynorm, satisfactionnorm)
word_embeddings <- word_embeddings_4
word_embeddings_wordnorm <- textEmbed(norms)
similarity_scores <- textSimilarityNorm(
  word_embeddings$harmonytext,
  word_embeddings_wordnorm$harmonynorm
)
}
}
\seealso{
see \code{\link{textSimilarity}} and \code{\link{textSimilarityTest}}
}
