% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_2_textTopicsAnalysis.R
\name{textTopicsTest}
\alias{textTopicsTest}
\title{Wrapper for topicsTest function from the topics package}
\usage{
textTopicsTest(
  model,
  x_variable = NULL,
  y_variable = NULL,
  controls = c(),
  test_method = "default",
  p_adjust_method = "fdr",
  ...
)
}
\arguments{
\item{model}{(list) The trained model}

\item{x_variable}{(string) The x variable name to be predicted, and to be plotted (only needed for regression or correlation)}

\item{y_variable}{(string) The y variable name to be predicted, and to be plotted (only needed for regression or correlation)}

\item{controls}{(vector) The control variables (not supported yet)}

\item{test_method}{(string) The test method to use, either "correlation","t-test", "linear_regression","logistic_regression", or "ridge_regression"}

\item{p_adjust_method}{(character) Method to adjust/correct p-values for multiple comparisons
(default = "none"; see also "holm", "hochberg", "hommel", "bonferroni", "BH", "BY",  "fdr").}

\item{...}{Parameter settings from topicsTest in the topics-package.}
}
\value{
A list of the test results, test method, and prediction variable
}
\description{
Wrapper for topicsTest function from the topics package
}
