% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_1_textTrain.R
\name{textTrain}
\alias{textTrain}
\title{Train word embeddings to a numeric variable.}
\usage{
textTrain(
  x,
  y,
  nrFolds_k = 10,
  preProcessPCAthresh = 0.95,
  strata_y = "y",
  methodCor = "pearson",
  model_description = "Consider writing a description here"
)
}
\arguments{
\item{x}{Wordembeddings from textEmbed (or textLayerAggregation).}

\item{y}{Numeric variable to predict.}

\item{nrFolds_k}{Number of folds to use.}

\item{preProcessPCAthresh}{Preprocessing threshold for amount of variance to retain (default 0.95).}

\item{strata_y}{Variable to stratify according (default y; can set to NULL).}

\item{methodCor}{Type of correlation used in evaluation (default "pearson"; can set to "spearman" or "kendall").}

\item{model_description}{Text to describe your model (optional; good when sharing the model with others).}
}
\value{
A correlation between predicted and observed values; as well as a tibble of predicted values.
}
\description{
Train word embeddings to a numeric variable.
}
\examples{
wordembeddings <- wordembeddings4_10
ratings_data <- Language_based_assessment_data_8_10
wordembeddings <- textTrain(wordembeddings$harmonytext, ratings_data$hilstotal,
nrFolds_k = 2, strata_y = NULL)
}
\seealso{
see \code{\link{textLayerAggregation}} \code{\link{textTrainLists}}
\code{\link{textTrainMultiTexts}} \code{\link{textTrainRandomForest}} \code{\link{textDiff}}
}
