% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_textTrain.R
\name{textTrain}
\alias{textTrain}
\title{textTrain trains word embeddings to a numeric variable.}
\usage{
textTrain(
  x,
  y,
  nrFolds_k = 10,
  methodTrain = "ridge",
  preProcessTrain = "pca",
  preProcessThresh = 0.95,
  methodCor = "pearson",
  describe_model = "Describe the model and share it with others",
  ...
)
}
\arguments{
\item{x}{Wordembeddings from textImport.}

\item{y}{A numeric variables to predict.}

\item{nrFolds_k}{Number of folds to use.}

\item{methodTrain}{Method used to train representations.}

\item{preProcessTrain}{Preprocessing training.}

\item{preProcessThresh}{Preprocessing threshold.}

\item{methodCor}{Type of correlation used in evaluation.}

\item{describe_model}{Input text to describe your model.}

\item{...}{Arguments from caret::createFolds, caret::trainControl and caret::train.}
}
\value{
A correlation between predicted and observed values; as well as predicted values.
}
\description{
textTrain trains word embeddings to a numeric variable.
}
\examples{
wordembeddings <- wordembeddings4_10
ratings_data <- sq_data_tutorial4_10
wordembeddings <- textTrain(wordembeddings$harmonytext, ratings_data$hilstotal)
}
\seealso{
see \code{\link{textTtest}}
}
