% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_1_textTrain.R
\name{textTrain}
\alias{textTrain}
\title{textTrain trains word embeddings to a numeric variable.}
\usage{
textTrain(
  x,
  y,
  nrFolds_k = 10,
  preProcessPCAthresh = 0.95,
  strata_y = "y",
  methodCor = "pearson",
  describe_model = "Describe the model further and share it with others"
)
}
\arguments{
\item{x}{Wordembeddings from textImport.}

\item{y}{The numeric variable to predict.}

\item{nrFolds_k}{Number of folds to use.}

\item{preProcessPCAthresh}{Preprocessing threshold.}

\item{strata_y}{variables to stratify according; default y, can set to NULL}

\item{methodCor}{Type of correlation used in evaluation; default pearson (see also "spearman", "kendall").}

\item{describe_model}{Input text to describe your model.}
}
\value{
A correlation between predicted and observed values; as well as predicted values.
}
\description{
textTrain trains word embeddings to a numeric variable.
}
\examples{
wordembeddings <- wordembeddings4_10
ratings_data <- sq_data_tutorial4_10
wordembeddings <- textTrain(wordembeddings$harmonytext, ratings_data$hilstotal,
nrFolds_k = 2, strata_y = NULL)
}
\seealso{
see \code{\link{textTrainLists}} \code{\link{textDiff}}
}
