% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_5_textTrainPredictExamples.R
\name{textTrainExamples}
\alias{textTrainExamples}
\alias{textPredictExamples}
\title{Show language examples (Experimental)}
\usage{
textTrainExamples(
  text,
  target,
  predictions,
  n_examples = 5,
  selection_method = "min_max",
  selection_variable = "predictions",
  filter_words = NULL,
  target_color = "darkgreen",
  predictions_color = "darkblue",
  error_color = "darkred",
  distribution_color = c("darkgreen", "gray", "darkred"),
  figure_format = "svg",
  scatter_legend_dot_size = 4,
  scatter_legend_bg_dot_size = 2,
  scatter_legend_n = c(3, 3, 3),
  scatter_show_axis_values = TRUE,
  grid_legend_x_axes_label = "x",
  grid_legend_y_axes_label = "y",
  seed = 42
)

textPredictExamples(
  text,
  target,
  predictions,
  n_examples = 5,
  selection_method = "min_max",
  selection_variable = "predictions",
  filter_words = NULL,
  target_color = "darkgreen",
  predictions_color = "darkblue",
  error_color = "darkred",
  distribution_color = c("darkgreen", "gray", "darkred"),
  figure_format = "svg",
  scatter_legend_dot_size = 4,
  scatter_legend_bg_dot_size = 2,
  scatter_legend_n = c(3, 3, 3),
  scatter_show_axis_values = TRUE,
  grid_legend_x_axes_label = "x",
  grid_legend_y_axes_label = "y",
  seed = 42
)
}
\arguments{
\item{text}{(string) the language that was used for prediction/assessment/classification.}

\item{target}{(numeric) the variable used for training (y).}

\item{predictions}{(numeric) the outcome from the model (i.e., y_hat).}

\item{n_examples}{(integer) the number of language examples to show.}

\item{selection_method}{(string) specification of which examples to select "min", "max", "min_max", "min_mean_max", "quintiles".}

\item{selection_variable}{(string) selection method: "predictions", "error", or "targets".}

\item{filter_words}{(character vector) words required to be included in examples.}

\item{target_color}{(string)}

\item{predictions_color}{(string) = "darkblue",}

\item{error_color}{=  (string) "darkred",}

\item{distribution_color}{(string) colors of the distribution plot}

\item{figure_format}{(string) file format of the figures.}

\item{scatter_legend_dot_size}{(integer) The size of dots in the scatter legend.}

\item{scatter_legend_bg_dot_size}{(integer) The size of background dots in the scatter legend.}

\item{scatter_legend_n}{(numeric or vector) A vector determining the number of dots to emphasize in each quadrant of the scatter legend.
For example: c(1,0,1) result in one dot in each quadrant except for the middle quadrant.}

\item{scatter_show_axis_values}{(boolean) If TRUE, the estimate values are shown on the distribution plot axes.}

\item{grid_legend_x_axes_label}{x-axis label of the grid topic plot.}

\item{grid_legend_y_axes_label}{y-axis label of the grid topic plot.}

\item{seed}{(integer) The seed to set for reproducibility.}
}
\value{
A tibble including examples with descriptive variables.
}
\description{
This function selects language examples that been used in the textTrain() or textAssess() functions.
}
