% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_5_textTrainPredictExamples.R
\name{textTrainExamples}
\alias{textTrainExamples}
\alias{textPredictExamples}
\title{Show language examples (Experimental)}
\usage{
textTrainExamples(
  text,
  target,
  predictions,
  n_examples = 5,
  selection_method = "min_max",
  selection_variable = "predictions",
  include_words = NULL,
  target_color = "darkgreen",
  predictions_color = "darkblue",
  error_color = "darkred",
  distribution_color = c("darkgreen", "gray", "darkred"),
  figure_format = "svg"
)

textPredictExamples(
  text,
  target,
  predictions,
  n_examples = 5,
  selection_method = "min_max",
  selection_variable = "predictions",
  include_words = NULL,
  target_color = "darkgreen",
  predictions_color = "darkblue",
  error_color = "darkred",
  distribution_color = c("darkgreen", "gray", "darkred"),
  figure_format = "svg"
)
}
\arguments{
\item{text}{(string) the language that was used for prediction/assessment/classification.}

\item{target}{(numeric) the variable used for training (y).}

\item{predictions}{(numeric) the outcome from the model (i.e., y_hat).}

\item{n_examples}{(integer) the number of language examples to show.}

\item{selection_method}{(string) specification of which examples to select "min", "max", "min_max", "min_mean_max", "quintiles".}

\item{selection_variable}{(string) selection method: "predictions", "error", or "targets".}

\item{include_words}{(character vector) words required to be included in examples.}

\item{target_color}{(string)}

\item{predictions_color}{(string)= "darkblue",}

\item{error_color}{=  (string)"darkred",}

\item{distribution_color}{(string) colors of the distribution plot}

\item{figure_format}{(string) file format of the figures.}
}
\value{
A tibble including examples with descriptive variables.
}
\description{
This function selects language examples that been used in the textTrain() or textAssess() functions.
}
