% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_1_textTrain.R
\name{textTrainLists}
\alias{textTrainLists}
\title{Individually trains word embeddings from several text variables to several numeric/categorical variables.}
\usage{
textTrainLists(x, y, force_train_method = "automatic", ...)
}
\arguments{
\item{x}{Word embeddings from textEmbed (or textLayerAggregation).}

\item{y}{Numeric variable to predict.}

\item{force_train_method}{default is automatic; see also "regression" and "random_forest".}

\item{...}{Arguments from textTrainRegression or textTrainRandomForest the textTrain function.}
}
\value{
Correlations between predicted and observed values.
}
\description{
Individually trains word embeddings from several text variables to several numeric/categorical variables.
}
\examples{
wordembeddings <- wordembeddings4[1:2]
ratings_data <- Language_based_assessment_data_8[5:6]
results <- textTrainLists(wordembeddings,
ratings_data,
multi_cores = FALSE)
}
\seealso{
see \code{\link{textTrain}}  \code{\link{textTrainRegression}}  \code{\link{textTrainRandomForest}}
}
