% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_1_textTrain.R
\name{textTrainLists}
\alias{textTrainLists}
\title{Individually trains word embeddings from several text variables to several numeric/categorical variables.}
\usage{
textTrainLists(
  x,
  y,
  force_train_method = "regression",
  preprocess_PCA_thresh = 0.95,
  method_cor = "pearson",
  model_description = "Consider writing a description here",
  multi_cores = TRUE,
  trees = 500,
  ...
)
}
\arguments{
\item{x}{Word embeddings from textEmbed (or textLayerAggregation).}

\item{y}{Numeric variable to predict.}

\item{force_train_method}{default is "regression"; see also "random_forest".}

\item{preprocess_PCA_thresh}{Pre-processing threshold for amount of variance to retain (default 0.95).}

\item{method_cor}{Type of correlation used in evaluation (default "pearson";
can set to "spearman" or "kendall").}

\item{model_description}{Text to describe your model (optional; good when sharing the model with others).}

\item{multi_cores}{If TRUE enables the use of multiple cores if computer/system allows for it (hence it can
make the analyses considerably faster to run).}

\item{trees}{number of trees if it is a categorical variable.}

\item{...}{Arguments for the textTrain function.}
}
\value{
Correlations between predicted and observed values.
}
\description{
Individually trains word embeddings from several text variables to several numeric/categorical variables.
}
\examples{
wordembeddings <- wordembeddings4[1:2]
ratings_data <- Language_based_assessment_data_8[5:6]
results <- textTrainLists(wordembeddings,
ratings_data,
multi_cores = FALSE)
}
\seealso{
see \code{\link{textTrain}}
}
