% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_1_textTrain.R
\name{textTrainLists}
\alias{textTrainLists}
\title{Trains word embeddings from several text variables to several numeric variables.}
\usage{
textTrainLists(
  x,
  y,
  trainMethod = "regression",
  nrFolds_k = 10,
  preProcessPCAthresh = 0.95,
  strata_y = "y",
  methodCor = "pearson",
  trees = 500,
  nrFolds_k_out = 10,
  ...
)
}
\arguments{
\item{x}{List of lists comprising several word embeddings from textEmbed
(NB need to remove any word embeddings from the decontextualised single words).}

\item{y}{Tibble with numeric variables to predict.}

\item{trainMethod}{Method to train wordembeddings (default "regression"; see also "randomForest").}

\item{nrFolds_k}{Number of folds to use.}

\item{preProcessPCAthresh}{Preprocessing threshold for amount of variance to retain (default 0.95).}

\item{strata_y}{Variable to stratify according (default y; can set to NULL).}

\item{methodCor}{Type of correlation used in evaluation (default "pearson"; can set to "spearman" or "kendall").}

\item{trees}{Number of trees used in random forest.}

\item{nrFolds_k_out}{Number of outer folds (only for textTrainCVpredictions textTrainCVpredictionsRF,
which is not yet fully implemented).}

\item{...}{Arguments for the textTrain function.}
}
\value{
Correlations between predicted and observed values.
}
\description{
Trains word embeddings from several text variables to several numeric variables.
}
\examples{
wordembeddings <- wordembeddings4_10[1:2]
ratings_data <- Language_based_assessment_data_8_10[5:6]
wordembeddings <- textTrainLists(wordembeddings, ratings_data, nrFolds_k = 2)
}
\seealso{
see \code{\link{textTrain}}
}
