% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_2_textTrainMulti.R
\name{textTrainMultiTexts}
\alias{textTrainMultiTexts}
\title{textTrainMultiTexts trains word embeddings from several text variables to a numeric variable.}
\usage{
textTrainMultiTexts(
  xlist,
  y,
  preProcessThresh = 0.95,
  nrFolds_k = 10,
  strata_y = "y",
  methodCor = "pearson",
  describe_model = "Describe the model further and share it with others"
)
}
\arguments{
\item{xlist}{list of word embeddings from textImport.}

\item{y}{The numeric variable to predict.}

\item{preProcessThresh}{Preprocessing threshold.}

\item{nrFolds_k}{Number of folds to use.}

\item{strata_y}{variables to stratify according; default y, can set to NULL}

\item{methodCor}{Type of correlation used in evaluation; default pearson (see also "spearman", "kendall").}

\item{describe_model}{Input text to describe your model.}
}
\value{
A correlation between predicted and observed values; as well as predicted values.
}
\description{
Concenate word embeddings from several text xariables to predict an outcome. The word embeddings for each text variables
go through seperate PCA, where the PCA components acre concenated and used in a ridge regression.
}
\examples{
wordembeddings <- wordembeddings4_10[1:4]
ratings_data <- sq_data_tutorial4_10$hilstotal
wordembeddings <- textTrainMultiTexts(wordembeddings, ratings_data, nrFolds_k = 2)
}
\seealso{
see \code{\link{textTrainLists}} \code{\link{textTtest}}
}
