% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_4_textTrainMulti.R
\name{textTrainMultiTexts}
\alias{textTrainMultiTexts}
\title{Combine several word embeddings from several text variables and train to a numeric variable.
Takes several text variables; applies separate PCAs to each text's word embeddings separately;
concatenate the PCA components to then be trained to predict one outcome with ridge regression.}
\usage{
textTrainMultiTexts(
  xlist,
  y,
  preProcessThresh = 0.95,
  nrFolds_k = 10,
  strata_y = "y",
  methodCor = "pearson",
  describe_model = "Describe the model further and share it with others"
)
}
\arguments{
\item{xlist}{List of word embeddings from textEmbed.}

\item{y}{Numeric variable to predict.}

\item{preProcessThresh}{Pre-processing threshold for amount of variance to retain (default 0.95).}

\item{nrFolds_k}{Number of folds to use.}

\item{strata_y}{Variable to stratify according (default y; can set to NULL).}

\item{methodCor}{Type of correlation used in evaluation (default "pearson"; can set to "spearman" or "kendall").}

\item{describe_model}{Text to describe your model.}
}
\value{
A correlation between predicted and observed values; as well as predicted values.
}
\description{
Concatenate word embeddings from several text variables to predict an outcome.
The word embeddings for each text variable go through separate PCAs, where the PCA components are
concatenated and used in a ridge regression.
}
\examples{
wordembeddings <- wordembeddings4_10[1:4]
ratings_data <- Language_based_assessment_data_8_10$hilstotal
results <- textTrainMultiTexts(wordembeddings, ratings_data, nrFolds_k = 2)
}
\seealso{
see \code{\link{textTrainLists}} \code{\link{textDiff}}
}
