% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_3_textTrainRandomForest.R
\name{textTrainRandomForest}
\alias{textTrainRandomForest}
\title{textTrainRandomForest trains word embeddings to a categorical variable using random forrest.}
\usage{
textTrainRandomForest(
  x,
  y,
  trees = 500,
  nrFolds_k = 10,
  strata_y = "y",
  describe_model = "Describe the model further and share it with others"
)
}
\arguments{
\item{x}{Wordembeddings from textImport.}

\item{y}{The categorical variable to predict.}

\item{trees}{Number of trees to use default 500).}

\item{nrFolds_k}{Number of folds to use (default 10).}

\item{strata_y}{variables to stratify according; default y, can set to NULL}

\item{describe_model}{Input text to describe your model.}
}
\value{
A correlation between predicted and observed values; as well as predicted values.
}
\description{
textTrainRandomForest trains word embeddings to a categorical variable using random forrest.
}
\examples{
wordembeddings <- wordembeddings4_10
ratings_data <- sq_data_tutorial4_10
wordembeddings <- textTrain(wordembeddings$harmonytext, ratings_data$hilstotal, nrFolds_k = 2)
}
\seealso{
see \code{\link{textTrainLists}} \code{\link{textTtest}}
}
