% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_3_textTrainRandomForest.R
\name{textTrainRandomForest}
\alias{textTrainRandomForest}
\title{Train word embeddings to a categorical variable using random forrest.}
\usage{
textTrainRandomForest(
  x,
  y,
  trees = 500,
  nrFolds_k = 10,
  strata_y = "y",
  describe_model = "Describe the model further and share it with others"
)
}
\arguments{
\item{x}{Wordembeddings from textEmbed.}

\item{y}{Categorical variable to predict.}

\item{trees}{Number of trees to use (default 500).}

\item{nrFolds_k}{Number of folds to use (default 10).}

\item{strata_y}{Variable to stratify according (default y; can set to NULL).}

\item{describe_model}{Text to describe your model.}
}
\value{
A correlation between predicted and observed values; as well as predicted values.
}
\description{
Train word embeddings to a categorical variable using random forrest.
}
\examples{
wordembeddings <- wordembeddings4_10
ratings_data <- Language_based_assessment_data_8_10
wordembeddings <- textTrain(wordembeddings$harmonytext, ratings_data$hilstotal, nrFolds_k = 2)
}
\seealso{
see \code{\link{textTrainLists}} \code{\link{textDiff}}
}
