% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_3_textTrainRandomForest.R
\name{textTrainRandomForest}
\alias{textTrainRandomForest}
\title{Train word embeddings to a categorical variable using random forrest.}
\usage{
textTrainRandomForest(
  x,
  y,
  preprocess_PCA_thresh = c(0.75, 0.85, 0.95),
  mtry = c(1, 5, 10, 15, 30, 40),
  min_n = c(1, 5, 10, 15, 30, 40),
  trees = c(1000, 1500),
  eval_measure = "roc_auc",
  model_description = "Consider writing a description of your model here",
  multi_cores = TRUE
)
}
\arguments{
\item{x}{Word embeddings from textEmbed.}

\item{y}{Categorical variable to predict.}

\item{mtry}{hyper parameter that may be tuned;  default:c(1, 20, 40),}

\item{min_n}{hyper parameter that may be tuned; default: c(1, 20, 40)}

\item{trees}{Number of trees to use (default 1000).}

\item{eval_measure}{Measure to evaluate the models in order to select the best hyperparameters default "roc_auc";
see also "accuracy", "bal_accuracy", "sens", "spec", "precision", "kappa", "f_measure".}

\item{model_description}{Text to describe your model (optional; good when sharing the model with others).}

\item{multi_cores}{If TRUE enables the use of multiple cores if computer/system allows for it (hence it can
make the analyses considerably faster to run).}
}
\value{
A list with roc_curve_data, roc_curve_plot, truth and predictions, final_model, model_description
chisq and fishers test as well as evaluation measures, e.g., including accuracy, f_meas and roc_auc (for details on
these measures see the yardstick r-package documentation).
}
\description{
Train word embeddings to a categorical variable using random forrest.
}
\examples{
wordembeddings <- wordembeddings4_10
example_categories <- as.factor(c(1, 2, 1, 2, 1, 2, 1, 2, 1, 2))
results <- textTrainRandomForest(wordembeddings$harmonywords,
                                 example_categories,
                                 trees = c(1000, 1500)
                                 mtry  = c(1),       # this is short because of testing
                                 min_n = c(1),       # this is short because of testing
                                 multi_cores = FALSE # This is FALSE due to CRAN testing.
)
}
\seealso{
see \code{\link{textTrainLists}} \code{\link{textDiff}}
}
