% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_2_textTrainRegression.R
\name{textTrainRegression}
\alias{textTrainRegression}
\title{Train word embeddings to a numeric variable.}
\usage{
textTrainRegression(
  x,
  y,
  model = "regression",
  eval_measure = "default",
  preprocess_PCA = NA,
  penalty = 10^seq(-16, 16),
  mixture = c(0),
  method_cor = "pearson",
  model_description = "Consider writing a description of your model here",
  multi_cores = FALSE,
  save_output = "all"
)
}
\arguments{
\item{x}{Word embeddings from textEmbed (or textEmbedLayerAggreation).}

\item{y}{Numeric variable to predict.}

\item{model}{Type of model. Default is "regression"; see also "logistic" for classification.}

\item{eval_measure}{Type of evaluative measure to select models from; default =  "rmse" for regression and "bal_accuracy"
for logistic.}

\item{preprocess_PCA}{Pre-processing threshold for PCA (to skip this step set it to NA).
Can select amount of variance to retain (e.g., .90 or as a grid c(0.80, 0.90)); or
number of components to select (e.g., 10). Default is "min_halving", which is a function
that selects the number of PCA components based on number  of participants and feature (word embedding dimensions)
in the data. The formula is:
preprocess_PCA = round(max(min(number_features/2), number_participants/2), min(50, number_features))).}

\item{penalty}{hyper parameter that is tuned}

\item{mixture}{hyper parameter that is tuned default = 0 (hence a pure ridge regression).}

\item{method_cor}{Type of correlation used in evaluation (default "pearson";
can set to "spearman" or "kendall").}

\item{model_description}{Text to describe your model (optional; good when sharing the model with others).}

\item{multi_cores}{If TRUE enables the use of multiple cores if computer/system allows for it (hence it can
make the analyses considerably faster to run; does not work on Windows).}

\item{save_output}{Option not to save all output; default "all". see also "only_results" and "only_results_predictions".}
}
\value{
A (one-sided) correlation test between predicted and observed values; tibble of predicted values, as well as information
about the model (preprossing_recipe, final_model and model_description).
}
\description{
Train word embeddings to a numeric variable.
}
\examples{
wordembeddings <- wordembeddings4
ratings_data <- Language_based_assessment_data_8

results <- textTrainRegression(wordembeddings$harmonytext,
  ratings_data$hilstotal,
  multi_cores = FALSE # this is set to FALSE due to CRAN testing, and Windows machines.
)
}
\seealso{
see \code{\link{textEmbedLayerAggreation}} \code{\link{textTrainLists}}
\code{\link{textTrainRandomForest}} \code{\link{textSimilarityTest}}
}
