% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_2_textTrainRegression.R
\name{textTrainRegression}
\alias{textTrainRegression}
\title{Train word embeddings to a numeric variable.}
\usage{
textTrainRegression(
  x,
  y,
  outside_strata_y = "y",
  inside_strata_y = "y",
  preprocess_PCA_thresh = 0.95,
  penalty = 10^seq(-16, 16),
  mixture = c(0),
  method_cor = "pearson",
  model_description = "Consider writing a description of your model here",
  multi_cores = TRUE
)
}
\arguments{
\item{x}{Word embeddings from textEmbed (or textLayerAggregation).}

\item{y}{Numeric variable to predict.}

\item{outside_strata_y}{Variable to stratify according (default y; can set to NULL).}

\item{inside_strata_y}{Variable to stratify according (default y; can set to NULL).}

\item{preprocess_PCA_thresh}{Pre-processing threshold for amount of variance to retain (default 0.95).}

\item{penalty}{hyper parameter that is tuned}

\item{mixture}{hyper parameter that is tuned default = 0 (hence a pure ridge regression).}

\item{method_cor}{Type of correlation used in evaluation (default "pearson";
can set to "spearman" or "kendall").}

\item{model_description}{Text to describe your model (optional; good when sharing the model with others).}

\item{multi_cores}{If TRUE enables the use of multiple cores if computer/system allows for it (hence it can
make the analyses considerably faster to run).}
}
\value{
A correlation between predicted and observed values; as well as a tibble of predicted values.
}
\description{
Train word embeddings to a numeric variable.
}
\examples{
wordembeddings <- wordembeddings4_10
ratings_data <- Language_based_assessment_data_8_10

results <- textTrainRegression(wordembeddings$harmonytext,
ratings_data$hilstotal,
outside_strata_y = NULL,
inside_strata_y = NULL,
multi_cores = FALSE #this is FALSE due to CRAN testing.
)
}
\seealso{
see \code{\link{textLayerAggregation}} \code{\link{textTrainLists}}
\code{\link{textTrainMultiTexts}} \code{\link{textTrainRandomForest}} \code{\link{textDiff}}
}
