% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_4_textPlotWordPrediction.R
\name{textWordPrediction}
\alias{textWordPrediction}
\title{Compute Supervised Dimension Projection and related variables for plotting words.}
\usage{
textWordPrediction(
  words,
  single_wordembeddings = single_wordembeddings_df,
  x,
  y = NULL,
  seed = 1003,
  case_insensitive = TRUE
)
}
\arguments{
\item{words}{Word or text variable to be plotted.}

\item{single_wordembeddings}{Word embeddings from textEmbed for individual words
(i.e., decontextualized embeddings).}

\item{x}{Numeric variable that the words should be plotted according to on the x-axes.}

\item{y}{Numeric variable that the words should be plotted according to on the y-axes (y=NULL).}

\item{seed}{Set different seed.}

\item{...}{Training options from textTrainRegression().}
}
\value{
A dataframe with variables (e.g., including Supervised Dimension Projection, frequencies, p-values)
for the individual words that is used for the plotting in the textProjectionPlot function.
}
\description{
Compute Supervised Dimension Projection and related variables for plotting words.
}
\examples{
# Data
wordembeddings <- wordembeddings4
raw_data <- Language_based_assessment_data_8
# Pre-processing data for plotting
df_for_plotting <- textProjection(
  words = raw_data$harmonywords,
  wordembeddings = wordembeddings$harmonywords,
  single_wordembeddings = wordembeddings$singlewords_we,
  x = raw_data$hilstotal,
  split = "mean",
  Npermutations = 10,
  n_per_split = 1
)
df_for_plotting
#' @seealso see \code{\link{textProjection}}
}
